% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{class}
\name{alg-class}
\alias{alg-class}
\title{Abstract optimization algorithm class}
\description{
An S4 class to represent an abstract optimization algorithm.
}
\section{Slots}{

\describe{
\item{\code{name}}{algorithm name}
}}

