% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{class}
\name{algs-classes}
\alias{algs-classes}
\alias{alg-class}
\alias{multistarts-class}
\alias{seed-class}
\alias{hybrid-class}
\alias{genetic-class}
\title{Optimization algorithm classes}
\description{
Optimization algorithm classes

alg 
An S4 class to represent an abstract optimisation algorithm.

gree 
An S4 class to represent a greedy algorithm extends \code{alg} class with multiple start.

seed 
An S4 class to represent a greedy algorithm extends \code{alg} class with initialization with spectral clustering and or k-means.

genetic
An S4 class to represent a hybrid genetic/greedy algorithm extends \code{alg} class.

genetic
An S4 class to represent a hybrid genetic/greedy algorithm extends \code{alg} class.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the algorithm}

\item{\code{nb_start}}{number of random starts (default to 10)}

\item{\code{pop_size}}{size of the solutions populations (default to 10)}

\item{\code{nb_max_gen}}{maximal number of generation to produce (default to 4)}

\item{\code{pop_size}}{size of the solutions populations (default to 10)}

\item{\code{nb_max_gen}}{maximal number of generation to produce (default to 4)}
}}

