% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_misc.R
\name{available_algorithms}
\alias{available_algorithms}
\title{Display the list of every currently available optimization algorithm}
\usage{
available_algorithms()
}
\description{
Display the list of every currently available optimization algorithm
}
