% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{clustering}
\alias{clustering}
\title{Generic method to extract the clustering results from an \code{\link{IclFit-class}} object}
\usage{
clustering(fit)
}
\arguments{
\item{fit}{an \code{IclFit} solution}
}
\value{
an integer vector with cluster assignments. Zero indicates noise points.
}
\description{
This method take a \code{\link{IclFit-class}} object and return an integer vector with the cluster assignments that were found.
}
