% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{clustplot,mm_fit-method}
\alias{clustplot,mm_fit-method}
\title{Plot a clustering results}
\usage{
\S4method{clustplot}{mm_fit}(sol)
}
\arguments{
\item{sol}{\code{\link{mm_fit}} obejct to be ploted}
}
\value{
a ggplot2 graphics which summarize the results
}
\description{
Plot a clustering results
}
