% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{clustplot,sbm_fit-method}
\alias{clustplot,sbm_fit-method}
\title{Plot a clustering results}
\usage{
\S4method{clustplot}{sbm_fit}(sol)
}
\arguments{
\item{sol}{\code{\link{sbm_fit}} obejct to be ploted}
}
\value{
a ggplot2 graphics which summarize the results
}
\description{
Plot a clustering results
}
