% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clustplot}
\alias{clustplot}
\title{Plot a clustering results}
\usage{
clustplot(sol)
}
\arguments{
\item{sol}{\code{\link{icl_fit}} obejct to be ploted}
}
\value{
a ggplot2 graphics which summarize the results
}
\description{
Plot a clustering results
}
