% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclust_dcsbm.R
\name{coef,co_dcsbm_fit-method}
\alias{coef,co_dcsbm_fit-method}
\title{Extract parameters from an \code{\link{co_dcsbm_fit-class}} object}
\usage{
\S4method{coef}{co_dcsbm_fit}(object)
}
\arguments{
\item{object}{a \code{\link{co_dcsbm_fit-class}}}
}
\value{
a list with the model parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pirows'}: row cluster proportions 
\item \code{'picols'}: row cluster proportions 
\item \code{'thetakl'}: between clusters connection probabilities (matrix of size Krow x Kcol),
\item \code{'gammarows'}: rows degree correction parameters (size Nrows),
\item \code{'gammacols'}: cols degree correction parameters (size Ncols),
}
}
\description{
Extract parameters from an \code{\link{co_dcsbm_fit-class}} object
}
