% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsbm.R
\name{coef,dcsbm_fit-method}
\alias{coef,dcsbm_fit-method}
\title{Extract parameters from an \code{\link{dcsbm_fit-class}} object}
\usage{
\S4method{coef}{dcsbm_fit}(object)
}
\arguments{
\item{object}{a \code{\link{dcsbm_fit-class}}}
}
\value{
a list with the model parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions 
\item \code{'thetakl'}: beetween clusters connections normalized intensities (matrix of size K x K), 
}
}
\description{
Extract parameters from an \code{\link{dcsbm_fit-class}} object
}
\details{
Currently the parameter \code{thetakl} is estimated by MLE.
}
