% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dcsbm.R
\name{coef,DcSbmFit-method}
\alias{coef,DcSbmFit-method}
\title{Extract parameters from an \code{\link{DcSbmFit-class}} object}
\usage{
\S4method{coef}{DcSbmFit}(object)
}
\arguments{
\item{object}{a \code{\link{DcSbmFit-class}}}
}
\value{
a list with the model parameters estimates (MAP), the fields are the following for "directed" models :
\itemize{
\item \code{'pi'}: cluster proportions
\item \code{'thetakl'}: between cluster normalized connection intensities (matrix of size K x K),
\item \code{gammain}: node in-degree correction parameter
\item \code{gammaout}: node out-degree correction parameter
}
And as follow for un-directed models :
#' \itemize{
\item \code{'pi'}: cluster proportions
\item \code{'thetakl'}: between cluster normalized connection intensities (matrix of size K x K),
\item \code{gamma}: node degree correction parameter
}
}
\description{
Extract parameters from an \code{\link{DcSbmFit-class}} object
}
\details{
in case of undirected graph
}
