% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diaggmm.R
\name{coef,diaggmm_fit-method}
\alias{coef,diaggmm_fit-method}
\title{Extract mixture parameters from \code{\link{diaggmm_fit-class}} object}
\usage{
\S4method{coef}{diaggmm_fit}(object)
}
\arguments{
\item{object}{a \code{\link{diaggmm_fit-class}}}
}
\value{
a list with the mixture parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions 
\item \code{'muk'}: cluster means
\item \code{'Sigmak'}: cluster co-variance matrices
}
}
\description{
Extract mixture parameters from \code{\link{diaggmm_fit-class}} object
}
