% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diaggmm.R
\name{coef,DiagGmmFit-method}
\alias{coef,DiagGmmFit-method}
\title{Extract mixture parameters from \code{\link{DiagGmmFit-class}} object}
\usage{
\S4method{coef}{DiagGmmFit}(object)
}
\arguments{
\item{object}{a \code{\link{DiagGmmFit-class}}}
}
\value{
a list with the mixture parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions
\item \code{'muk'}: cluster means
\item \code{'Sigmak'}: cluster co-variance matrices
}
}
\description{
Extract mixture parameters from \code{\link{DiagGmmFit-class}} object
}
