% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gmm.R
\name{coef,GmmFit-method}
\alias{coef,GmmFit-method}
\title{Extract mixture parameters from \code{\link{GmmFit-class}} object}
\usage{
\S4method{coef}{GmmFit}(object)
}
\arguments{
\item{object}{a \code{\link{GmmFit-class}}}
}
\value{
a list with the mixture parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions
\item \code{'muk'}: cluster means
\item \code{'Sigmak'}: cluster co-variance matrices
}
}
\description{
Extract mixture parameters from \code{\link{GmmFit-class}} object
}
