% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{coef,IclFit-method}
\alias{coef,IclFit-method}
\title{Extract parameters from an \code{\link{IclFit-class}} object}
\usage{
\S4method{coef}{IclFit}(object)
}
\arguments{
\item{object}{a \code{\link{IclFit-class}}}
}
\value{
a list with the model parameters estimates (MAP)
}
\description{
Extract parameters from an \code{\link{IclFit-class}} object
}
\details{
The results depends of the used model, in case the method is not yet implemented for a model, this generic method will be used. Which will return the \code{obs_stats} slot of the model.
}
