% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lca.R
\name{coef,LcaFit-method}
\alias{coef,LcaFit-method}
\title{Extract parameters from an \code{\link{LcaFit-class}} object}
\usage{
\S4method{coef}{LcaFit}(object)
}
\arguments{
\item{object}{a \code{\link{LcaFit-class}}}
}
\value{
a list with the model parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions
\item \code{'thetav'}: cluster profile probabilities (list of matrix of size K x Dv),
}
}
\description{
Extract parameters from an \code{\link{LcaFit-class}} object
}
