% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misssbm.R
\name{coef,misssbm_fit-method}
\alias{coef,misssbm_fit-method}
\title{Extract parameters from an \code{\link{misssbm_fit-class}} object}
\usage{
\S4method{coef}{misssbm_fit}(object)
}
\arguments{
\item{object}{a \code{\link{misssbm_fit-class}}}
}
\value{
a list with the model parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions 
\item \code{'thetakl'}: between clusters connection probabilites (array of size K x K),
\item \code{'epsilonkl'}: between clusters dyad observation probabilites (array of size K x K) for block-dyad sampling and double for dyad sampling,
}
}
\description{
Extract parameters from an \code{\link{misssbm_fit-class}} object
}
