% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{coef,mm_fit-method}
\alias{coef,mm_fit-method}
\title{Extract parameters from an \code{\link{mm_fit-class}} object}
\usage{
\S4method{coef}{mm_fit}(object)
}
\arguments{
\item{object}{a \code{\link{mm_fit-class}}}
}
\value{
a list with the model parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions 
\item \code{'thetak'}: cluster profile probabilites (matrix of size K x D), 
}
}
\description{
Extract parameters from an \code{\link{mm_fit-class}} object
}
