% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mmm.R
\name{coef,mmm_fit-method}
\alias{coef,mmm_fit-method}
\title{Extract parameters from an \code{\link{mmm_fit-class}} object}
\usage{
\S4method{coef}{mmm_fit}(object)
}
\arguments{
\item{object}{a \code{\link{mmm_fit-class}}}
}
\value{
a list with the model parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions 
\item \code{'thetav'}: cluster profile probabilites (list of matrix of size K x Dv), 
}
}
\description{
Extract parameters from an \code{\link{mmm_fit-class}} object
}
