% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mom.R
\name{coef,MoMFit-method}
\alias{coef,MoMFit-method}
\title{Extract parameters from an \code{\link{MoMFit-class}} object}
\usage{
\S4method{coef}{MoMFit}(object)
}
\arguments{
\item{object}{a \code{\link{MoMFit-class}}}
}
\value{
a list with the model parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions
\item \code{'thetak'}: cluster profile probabilities (matrix of size K x D),
}
}
\description{
Extract parameters from an \code{\link{MoMFit-class}} object
}
