% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mor.R
\name{coef,MoRFit-method}
\alias{coef,MoRFit-method}
\title{Extract mixture parameters from \code{\link{MoRFit-class}} object using MAP estimation}
\usage{
\S4method{coef}{MoRFit}(object)
}
\arguments{
\item{object}{a \code{\link{MoRFit-class}}}
}
\value{
a list with the mixture parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions
\item \code{'A'}: cluster regression matrix
\item \code{'Sigmak'}: cluster noise co-variance matrices
}
}
\description{
Extract mixture parameters from \code{\link{MoRFit-class}} object using MAP estimation
}
