% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multsbm.R
\name{coef,multsbm_fit-method}
\alias{coef,multsbm_fit-method}
\title{Extract parameters from an \code{\link{multsbm_fit-class}} object}
\usage{
\S4method{coef}{multsbm_fit}(object)
}
\arguments{
\item{object}{a \code{\link{multsbm_fit-class}}}
}
\value{
a list with the model parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions 
\item \code{'thetakl'}: cluster profile probabilities (array of size K x K x D), 
}
}
\description{
Extract parameters from an \code{\link{multsbm_fit-class}} object
}
