% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_multsbm.R
\name{coef,MultSbmFit-method}
\alias{coef,MultSbmFit-method}
\title{Extract parameters from an \code{\link{MultSbmFit-class}} object}
\usage{
\S4method{coef}{MultSbmFit}(object)
}
\arguments{
\item{object}{a \code{\link{MultSbmFit-class}}}
}
\value{
a list with the model parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions
\item \code{'thetakl'}: cluster profile probabilities (array of size K x K x D),
}
}
\description{
Extract parameters from an \code{\link{MultSbmFit-class}} object
}
