% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmreg.R
\name{coef,mvmreg_fit-method}
\alias{coef,mvmreg_fit-method}
\title{Extract mixture parameters from \code{\link{mvmreg_fit-class}} object}
\usage{
\S4method{coef}{mvmreg_fit}(object)
}
\arguments{
\item{object}{a \code{\link{mvmreg_fit-class}}}
}
\value{
a list with the mixture parameters estimates (MAP), the fields are:
\itemize{
\item \code{'pi'}: cluster proportions 
\item \code{'A'}: cluster regression matrix
\item \code{'Sigmak'}: cluster noise co-variance matrices
}
}
\description{
Extract mixture parameters from \code{\link{mvmreg_fit-class}} object
}
