% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_combinedmodels.R
\docType{class}
\name{CombinedModelsFit-class}
\alias{CombinedModelsFit-class}
\title{Combined Models fit results class}
\description{
An S4 class to represent a fit of a degree corrected stochastic block model for co_clustering, extend \code{\link{IclFit-class}}.
}
\section{Slots}{

\describe{
\item{\code{model}}{a \code{\link{DcSbm-class}} object to store the model fitted}

\item{\code{name}}{generative model name}

\item{\code{icl}}{icl value of the fitted model}

\item{\code{K}}{number of extracted clusters over row and columns}

\item{\code{cl}}{a numeric vector with row and columns cluster indexes}

\item{\code{obs_stats}}{a list with the following elements:}

\item{\code{move_mat}}{binary matrix which store move constraints}

\item{\code{train_hist}}{data.frame with training history information (details depends on the training procedure)}
}}

\seealso{
\code{\link{extractSubModel,CombinedModelsPath,character-method}}
}
