% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclust_dcsbm.R
\name{cut,co_dcsbm_path-method}
\alias{cut,co_dcsbm_path-method}
\title{method to cut a path solution to a desired number of cluster}
\usage{
\S4method{cut}{co_dcsbm_path}(x, K)
}
\arguments{
\item{x}{A an \code{\link{co_dcsbm_path-class}} solution}

\item{K}{Desired number of cluster}
}
\value{
an icl_path object with the desired number of cluster
}
\description{
this method take a \code{\link{co_dcsbm_path-class}} and an integer K and return the solution from the path with K clusters
}
