% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dclbm.R
\name{cut,DcLbmPath-method}
\alias{cut,DcLbmPath-method}
\title{Method to cut a DcLbmPath solution to a desired number of cluster}
\usage{
\S4method{cut}{DcLbmPath}(x, K)
}
\arguments{
\item{x}{A an \code{\link{DcLbmPath-class}} solution}

\item{K}{Desired number of cluster}
}
\value{
an \code{\link{IclPath-class}} object with the desired number of cluster
}
\description{
This method take a \code{\link{DcLbmPath-class}} and an integer K and return the solution from the path with K clusters
}
