% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{methods}
\name{cut,icl_path-method}
\alias{cut,icl_path-method}
\title{Cut a path to a desired number of cluster}
\usage{
\S4method{cut}{icl_path}(x, K)
}
\arguments{
\item{x}{A an \code{icl_path} solution}

\item{K}{Desired number of cluster}
}
\description{
Cut a path to a desired number of cluster
}
