% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{cut,IclPath-method}
\alias{cut,IclPath-method}
\title{Generic method to cut a path solution to a desired number of cluster}
\usage{
\S4method{cut}{IclPath}(x, K)
}
\arguments{
\item{x}{A an \code{IclPath} solution}

\item{K}{Desired number of cluster}
}
\value{
an \code{\link{IclPath-class}} object with the desired number of cluster
}
\description{
This method take a \code{\link{IclPath-class}} object and an integer K and return the solution from the path with K clusters
}
