% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{methods}
\name{cut,icl_path-method}
\alias{cut,icl_path-method}
\title{Cut
method to cut a path solution to a desired number of cluster}
\usage{
\S4method{cut}{icl_path}(x, K)
}
\arguments{
\item{x}{A an \code{icl_path} solution}

\item{K}{Desired number of cluster}
}
\value{
an icl_path obejct with the desired number of cluster
}
\description{
Cut
method to cut a path solution to a desired number of cluster
}
\section{Methods (by class)}{
\itemize{
\item \code{icl_path}: 
}}

