% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsbm.R
\docType{class}
\name{dcsbm-class}
\alias{dcsbm-class}
\title{Clustering with a degree corrected stochastic block model description class}
\description{
An S4 class to represent a degree corrected stochastic block model, extend \code{\link{icl_model-class}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter}

\item{\code{type}}{define the type of networks (either "directed" or "undirected", default to "directed")}
}}

