% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsbm.R
\docType{class}
\name{dcsbm_fit-class}
\alias{dcsbm_fit-class}
\title{Degree Corrected Stochastic Block Model fit results class}
\description{
An S4 class to represent a fit of a degree corrected stochastic block model for co_clustering, extend \code{\link{icl_fit-class}}.
}
\section{Slots}{

\describe{
\item{\code{model}}{a \code{\link{dcsbm-class}} object to store the model fitted}

\item{\code{name}}{generative model name}

\item{\code{icl}}{icl value of the fitted model}

\item{\code{K}}{number of extracted clusters over row and columns}

\item{\code{cl}}{a numeric vector with row and columns cluster indexes}

\item{\code{obs_stats}}{a list with the following elements:
\itemize{
\item counts: numeric vector of size K with number of elements in each clusters
\item din: numeric vector of size K which store the sums of in-degrees for each clusters
\item dout: numeric vector of size K which store the sums of out-degrees for each clusters 
\item x_counts: matrix of size K*K with the number of links between each pair of clusters 
}}

\item{\code{obs_stats_cst}}{a list with the following elements:
\itemize{
\item din_node: node in-degree, a vector of size N
\item dout_node: node in-degree vector of size N
}}

\item{\code{move_mat}}{binary matrix which store move constraints}

\item{\code{train_hist}}{data.frame with training history information (details depends on the training procedure)}
}}

