% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dcsbm.R
\docType{class}
\name{DcSbmPrior-class}
\alias{DcSbmPrior-class}
\alias{DcSbmPrior}
\alias{DcSbm-class}
\alias{DcSbm}
\title{Degree Corrected Stochastic Block Model Prior class}
\usage{
DcSbmPrior(p = NaN, type = "guess")

DcSbm(alpha = 1, p = NaN, type = "guess")
}
\description{
An S4 class to represent a degree corrected stochastic block model.
Such model can be used to cluster graph vertex, and model a square adjacency matrix \eqn{X} with the following generative model :
\deqn{ \pi \sim Dirichlet(\alpha)}
\deqn{ Z_i  \sim \mathcal{M}(1,\pi)}
\deqn{ \theta_{kl} \sim Exponential(p)}
\deqn{ \gamma_i^+,\gamma_i^- \sim \mathcal{U}(S_k)}
\deqn{ X_{ij}|Z_{ik}Z_{jl}=1 \sim \mathcal{P}(\gamma_i^+\theta_{kl}\gamma_j^-)}
The individuals parameters \eqn{\gamma_i^+,\gamma_i^-} allow to take into account the node degree heterogeneity.
These parameters have uniform priors over the simplex \eqn{S_k} ie. \eqn{\sum_{i:z_{ik}=1}\gamma_i^+=1}. This class mainly store the prior parameters value \eqn{\alpha} of this generative model in the following slots (the prior parameter \eqn{p} is estimated from the data as the global average probability of connection between two nodes):
}
\section{Functions}{
\itemize{
\item \code{DcSbmPrior}: DcSbmPrior class constructor

\item \code{DcSbm-class}: DcSbm class constructor

\item \code{DcSbm}: DcSbm class constructor
}}

\section{Slots}{

\describe{
\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter (default to 10)}

\item{\code{p}}{Exponential prior parameter (default to NaN, in this case p will be estimated from data as the mean connection probability)}

\item{\code{type}}{define the type of networks (either "directed", "undirected" or "guess", default to "guess")}
}}

\examples{
DcSbmPrior()
DcSbmPrior(type = "undirected")
DcSbm()
DcSbm(type = "undirected")
}
\seealso{
Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{GmmPrior-class}},
\code{\link{LcaPrior-class}},
\code{\link{MixedModels-class}},
\code{\link{MoMPrior-class}},
\code{\link{MoRPrior-class}},
\code{\link{MultSbmPrior-class}},
\code{\link{SbmPrior-class}}
}
\concept{DlvmModels}
