% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_classes.R
\docType{class}
\name{DlvmCoPrior-class}
\alias{DlvmCoPrior-class}
\title{Abstract class to represent a generative model for co-clustering}
\description{
An S4 class to represent an abstract generative model
}
\section{Slots}{

\describe{
\item{\code{alpha}}{a numeric vector of length 1 which define the parameters of the Dirichlet over the cluster proportions (default to 1)}
}}

\concept{DlvmCoModels}
