% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_classes.R
\docType{class}
\name{DlvmPrior-class}
\alias{DlvmPrior-class}
\title{abstract class to represent a generative model

An S4 class to represent an abstract generative model}
\description{
abstract class to represent a generative model

An S4 class to represent an abstract generative model
}
\section{Slots}{

\describe{
\item{\code{alpha}}{a numeric vector of length 1 which define the parameters of the Dirichlet over the cluster proportions (default to 1)}
}}

\seealso{
Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DcSbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{GmmPrior-class}},
\code{\link{LcaPrior-class}},
\code{\link{MixedModels-class}},
\code{\link{MoMPrior-class}},
\code{\link{MoRPrior-class}},
\code{\link{MultSbmPrior-class}},
\code{\link{SbmPrior-class}},
\code{\link{greed}()}
}
\concept{DlvmModels}
