% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mixedmodels.R
\name{extractSubModel}
\alias{extractSubModel}
\alias{extractSubModel,MixedModelsPath,character-method}
\title{Extract a part of a \code{\link{MixedModelsPath-class}} object}
\usage{
extractSubModel(sol, sub_model_name)

\S4method{extractSubModel}{MixedModelsPath,character}(sol, sub_model_name)
}
\arguments{
\item{sol}{an \code{\link{MixedModelsPath-class}} object}

\item{sub_model_name}{a string which specify the part of the model to
extract. Note that the name must correspond to the one of the names used in
the list of models during the origin call to \code{\link{greed}}.}
}
\value{
a \code{\link{IclFit-class}} object of the relevant class
}
\description{
Extract a part of a \code{\link{MixedModelsPath-class}} object
}
\section{Methods (by class)}{
\itemize{
\item \code{sol = MixedModelsPath,sub_model_name = character}: MixedModelsPath method
}}

