% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fashion}
\alias{fashion}
\title{Fashion mnist dataset}
\format{
An object of class \code{matrix} with a random sample of 1000 images (one per rows) extracted from the fashionmnist dataset.
}
\source{
\href{https://github.com/zalandoresearch/fashion-mnist}{https://github.com/zalandoresearch/fashion-mnist}
}
\usage{
data(fashion)
}
\description{
Zalando fashionmnist dataset, sample of 1 000 Zalando's article images from the test set.
}
\examples{
data(fashion)
}
\references{
Fashion-MNIST: a Novel Image Dataset for Benchmarking Machine Learning Algorithms. Han Xiao, Kashif Rasul, Roland Vollgraf (2017)
(\href{https://arxiv.org/abs/1708.07747}{arXiv:1708.07747}).
}
\keyword{datasets}
