% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fifa_positions}
\alias{Fifa_positions}
\title{Fifa_positions data}
\format{
An object of class \code{list} with two fields; \describe{
  \item{bg_img}{a matrix containing raw pixel representation of an image of a football
  field (useful for plotting payers).}
  \item{positions}{An R data.frame with 36 rows and 3 columns describing position, x and y axis on the field respectively.}
  }
}
\usage{
data(Fifa)
}
\description{
A subset of the \code{\link{Fifa}} dataset with 36 possible players positions containing their
x and y coordinate on a 2D image of a football field. Mainly
useful for plotting.
}
\examples{
data(Fifa_positions)
}
\keyword{datasets}
