% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,dgCMatrix,numeric,missing,missing-method}
\alias{fit,dgCMatrix,numeric,icl_model,missing-method}
\alias{fit,dgCMatrix,numeric,icl_model,genetic-method}
\alias{fit,dgCMatrix,numeric,icl_model,greed-method}
\title{Fit a clustering model}
\usage{
fit(x, K, model, alg)

\S4method{fit}{dgCMatrix,numeric,missing,missing}(x, K)

\S4method{fit}{dgCMatrix,numeric,icl_model,missing}(x, K, model)

\S4method{fit}{dgCMatrix,numeric,icl_model,genetic}(x, K, model, alg)

\S4method{fit}{dgCMatrix,numeric,icl_model,greed}(x, K, model, alg)
}
\arguments{
\item{x}{A sparse Matrix dgCMatrix}

\item{K}{An initial guess of the maximal number of cluster}

\item{model}{An IclModel}

\item{alg}{An optimization algorithm}

\item{x}{A sparse Matrix dgCMatrix}

\item{K}{An initial guess of the maximal number of cluster}

\item{x}{A sparse Matrix dgCMatrix}

\item{K}{An initial guess of the maximal number of cluster}

\item{model}{An IclModel}

\item{x}{A sparse Matrix dgCMatrix}

\item{K}{An initial guess of the maximal number of cluster}

\item{model}{An IclModel}

\item{alg}{An optimization algorithm}

\item{x}{A sparse Matrix dgCMatrix}

\item{K}{An initial guess of the maximal number of cluster}

\item{model}{An IclModel}

\item{alg}{An optimization algorithm}
}
\description{
Fit a clustering model

Fit a clustering model

Fit a clustering model

Fit a clustering model

Fit a clustering model
}
\section{Methods (by class)}{
\itemize{
\item \code{x = dgCMatrix,K = numeric,model = missing,alg = missing}: 

\item \code{x = dgCMatrix,K = numeric,model = icl_model,alg = missing}: 

\item \code{x = dgCMatrix,K = numeric,model = icl_model,alg = genetic}: 

\item \code{x = dgCMatrix,K = numeric,model = icl_model,alg = greed}: 
}}

