% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{methods}
\name{fit_cond,icl_model,hybrid-method}
\alias{fit_cond,icl_model,hybrid-method}
\title{#' @describeIn fit 
#' @export
setMethod(f = "fit", 
          signature = signature("icl_model","missing"), 
          definition = function(model,x, K=20,verbose=FALSE){
            fit(model,new("hybrid"),x,K,verbose)
          })}
\usage{
\S4method{fit_cond}{icl_model,hybrid}(model, alg, x, K, verbose = FALSE)
}
\description{
#' @describeIn fit 
#' @export
setMethod(f = "fit", 
          signature = signature("icl_model","missing"), 
          definition = function(model,x, K=20,verbose=FALSE){
            fit(model,new("hybrid"),x,K,verbose)
          })
}
