% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{methods}
\name{fit_cond}
\alias{fit_cond}
\alias{fit_cond,icl_model,seed-method}
\title{Fit a clustering model}
\usage{
fit_cond(model, alg, ...)

\S4method{fit_cond}{icl_model,seed}(model, alg, x, y, K = 20,
  verbose = FALSE)
}
\arguments{
\item{model}{An \code{\link{IclModel-class}} such as \code{\link{sbm-class}}, \code{\link{dcsbm-class}}, ...}

\item{alg}{An optimization algorithm such as \code{\link{greed-class}}, \code{\link{genetic-class}} or \code{\link{km-class}}}

\item{x}{A sparse Matrix as a \code{dgCMatrix}}

\item{K}{An initial guess of the maximal number of cluster}
}
\description{
Fit a clustering model
}
\section{Methods (by class)}{
\itemize{
\item \code{model = icl_model,alg = seed}: 
}}

