% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_greed_cond}
\alias{fit_greed_cond}
\title{fit_greed}
\usage{
fit_greed_cond(model, X, y, Ki, type = "both", nb_max_pass = 50L,
  verbose = FALSE)
}
\arguments{
\item{model}{icl_model}

\item{Ki}{initia guess for K}

\item{type}{: merge, swap or both (default)}

\item{nb_max_pass}{: maximum number of pass for greedy swap}

\item{xp}{sparseMatrix}
}
\description{
fit_greed
}
