% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_greed_init_type}
\alias{fit_greed_init_type}
\title{fit_greed_init_type}
\usage{
fit_greed_init_type(model, xp, Ki, clt, type)
}
\arguments{
\item{model}{icl_model}

\item{xp}{sparseMatrix}

\item{Ki}{initia guess for K}

\item{clt}{luster labels}
}
\description{
fit_greed_init_type
}
