% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_greed_path}
\alias{fit_greed_path}
\title{fit_greed_path}
\usage{
fit_greed_path(data, init_fit)
}
\arguments{
\item{data}{list with clustering data depnds on model type}

\item{init_fit}{initial fit object}
}
\value{
a model_path object
}
\description{
fit_greed_path
}
