% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_classes.R
\docType{class}
\name{fits-classes}
\alias{fits-classes}
\alias{icl_fit-class}
\alias{mm_fit-class}
\alias{mm_path-class}
\alias{sbm_path-class}
\alias{dcsbm_path-class}
\title{Clustering solutions classes}
\description{
Clustering solutions classes

icl_fit

An S4 abstract class to represent an icl fit of a clustering model.
\itemize{
\item slots : \code{name,K,icl,cl,count}
}

mm_fit

An S4 class to represent an icl fit of a mixture of multinomials model that extend \code{icl_fit}.
\itemize{
\item slots : \code{name,K,icl,cl,counts,x_counts,model}
}

mm_path

An S4 class to represent a hierachical path of stochastick block models that extend \code{icl_fit}.
\itemize{
\item slots : \code{name,K,icl,cl,counts,x_counts,path}
}

sbm_path

An S4 class to represent a hierachical path pf mixture of multinomials models that extend \code{icl_fit}.
\itemize{
\item slots : \code{name,K,icl,cl,counts,x_counts,path}
}

sbm_path

An S4 class to represent a hierachical path pf mixture of multinomials models that extend \code{icl_fit}.
\itemize{
\item slots : \code{name,K,icl,cl,counts,x_counts,path}
}
}
\section{Slots}{

\describe{
\item{\code{name}}{of the fit}

\item{\code{K}}{a numeric vector of length 1 which correspond to the number of clusters}

\item{\code{icl}}{a numeric vector of length 1 which store the the icl value}

\item{\code{cl}}{a numeric vector of length N which store the clusters labels}

\item{\code{counts}}{a numeric vector of length K which store the counts for each cluster}

\item{\code{path}}{a list of merge moves describing the hierachie of merge followed to complete totaly the merge path.}
}}

