% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_classes.R
\docType{class}
\name{sbm_fit-class}
\alias{sbm_fit-class}
\alias{dcsbm_fit-class}
\title{sbm_fit

An S4 class to represent a fit of a stochastick block model that extend \code{icl_fit}.
\itemize{
\item slots : \code{name,K,icl,cl,counts,x_counts,model}
}}
\description{
sbm_fit

An S4 class to represent a fit of a stochastick block model that extend \code{icl_fit}.
\itemize{
\item slots : \code{name,K,icl,cl,counts,x_counts,model}
}

dcsbm_fit

An S4 class to represent a fit of a stochastick block model that extend \code{icl_fit}.
\itemize{
\item slots : \code{name,K,icl,cl,counts,x_counts,model}
}
}
\section{Slots}{

\describe{
\item{\code{x_counts}}{a numeric matrix to store the observed counts in the data for each cluster}

\item{\code{model}}{an \code{\link{icl_model}} to store the model fitted}

\item{\code{x_counts}}{a numeric matrix to store the observed counts in the data for each cluster}

\item{\code{model}}{an \code{\link{icl_model}} to store the model fitted}
}}

