% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FrenchParliament}
\alias{FrenchParliament}
\title{French Parliament votes}
\format{An object of class \code{"list"} with two fields; 
 \describe{
  \item{X}{matrix of deputy votes a \cite{sparseMatrix} of size 593x570}
  \item{labels}{a data frame of deputy meta-data}
}}
\usage{
data(FrenchParliament)
}
\description{
French Parliament votes
}
\examples{
data(FrenchParliament)
}
\keyword{datasets}
