% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{graph_balance}
\alias{graph_balance}
\title{graph_balance}
\usage{
graph_balance(x)
}
\arguments{
\item{x}{a \code{\link{sbm_fit-class}} object to be plot}
}
\value{
a ggplot2 graph
}
\description{
graph_balance
}
