\name{greed-package}
\alias{greed-package}
\alias{greed}
\docType{package}
\title{
\packageTitle{greed}
}
\description{
\packageDescription{greed}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{greed}
\packageIndices{greed}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{greed}

Maintainer: \packageMaintainer{greed}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
