% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{greed_cond}
\alias{greed_cond}
\title{greed_cond}
\usage{
greed_cond(X, y, K = 20, model = find_model_cond(X, y),
  alg = new("hybrid"), verbose = FALSE)
}
\arguments{
\item{K}{Desired number of cluster}

\item{x}{An adjacency matrix in sparse format}
}
\description{
greed_cond
}
