% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_misc.R
\name{H}
\alias{H}
\title{Compute the entropy of a discrete sample}
\usage{
H(cl)
}
\arguments{
\item{cl}{vector of discrete labels}
}
\value{
the entropy of the sample
}
\description{
Compute the entropy of a discrete sample
}
\examples{
cl <- sample(2, 500, replace = TRUE)
H(cl)
}
