% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{class}
\name{hybrid-class}
\alias{hybrid-class}
\title{Hybrid optimization algorithm}
\description{
An S4 class to represent an hybrid genetic/greedy algorithm (extends \code{\link{alg-class}} class).
}
\section{Slots}{

\describe{
\item{\code{pop_size}}{size of the solutions populations (default to 20)}

\item{\code{nb_max_gen}}{maximal number of generation to produce (default to 10)}

\item{\code{prob_mutation}}{mutation probability (default to 0.25)}

\item{\code{Kmax}}{maximum number of clusters (default to 100)}
}}

