% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{ICL,IclFit-method}
\alias{ICL,IclFit-method}
\title{Method to extract the ICL value from an \code{\link{IclFit-class}} object}
\usage{
\S4method{ICL}{IclFit}(fit)
}
\arguments{
\item{x}{an \code{IclFit} solution}
}
\value{
The ICL value achieved
}
\description{
This method take a \code{\link{IclFit-class}} object and return its ICL score.
}
