% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{ICL}
\alias{ICL}
\alias{ICL,IclFit-method}
\title{Generic method to extract the ICL value from an \code{\link{IclFit-class}} object}
\usage{
ICL(fit)

\S4method{ICL}{IclFit}(fit)
}
\arguments{
\item{fit}{an \code{IclFit} solution}
}
\value{
The ICL value achieved
}
\description{
This method take a \code{\link{IclFit-class}} object and return its ICL score.
}
\section{Methods (by class)}{
\itemize{
\item \code{IclFit}: IclFit method
}}

