% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_classes.R
\docType{class}
\name{icl_fit-class}
\alias{icl_fit-class}
\title{abstract class to represent a clustering result}
\description{
An S4 abstract class to represent an icl fit of a clustering model.
}
\section{Slots}{

\describe{
\item{\code{K}}{a numeric vector of length 1 which correspond to the number of clusters}

\item{\code{icl}}{a numeric vector of length 1 which store the the icl value}

\item{\code{cl}}{a numeric vector of length N which store the clusters labels}

\item{\code{obs_stats}}{a list to store the observed statistics of the model needed to compute ICL.}

\item{\code{move_mat}}{binary matrix which store move constraints}

\item{\code{train_hist}}{a data.frame to store training history (format depends on the used algorithm used).}

\item{\code{name}}{generative model name}
}}

\seealso{
\code{\link{sbm_fit-class}}, \code{\link{dcsbm_fit-class}}, \code{\link{co_dcsbm_fit-class}}
}
