% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_classes.R
\docType{class}
\name{icl_model-class}
\alias{icl_model-class}
\title{abstract class to represent a generative model

An S4 class to represent an abstract generative model}
\description{
abstract class to represent a generative model

An S4 class to represent an abstract generative model
}
\section{Slots}{

\describe{
\item{\code{name}}{a character vector}

\item{\code{alpha}}{a numeric vector of length 1 which define the parameters of the Dirichlet over the cluster proportions (default to 1)}
}}

\seealso{
\code{\link{sbm-class}}, \code{\link{dcsbm-class}},\code{\link{co_dcsbm-class}},
}
