% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_cond}
\alias{init_cond}
\title{init}
\usage{
init_cond(model, X, y, clt)
}
\arguments{
\item{model}{icl_model}

\item{clt}{cluster labels {0,...,K-1}}

\item{xp}{sparseMatrix}
}
\description{
init
}
