% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Jazz}
\alias{Jazz}
\title{Jazz musicians network dataset}
\format{
An object of class \code{\link[Matrix]{sparseMatrix}} with the network adjacency matrix.
}
\usage{
data(Jazz)
}
\description{
List of edges of the network of Jazz musicians.
}
\examples{
data(Jazz)
}
\references{
P.Gleiser and L. Danon , Community Structure in jazz, Adv. Complex Syst.6, 565 (2003) (\href{https://arxiv.org/abs/cond-mat/0307434}{Arxiv})
}
\keyword{datasets}
