% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Jazz_full}
\alias{Jazz_full}
\title{Jazz musicians / Bands relations}
\format{
A list with two fields:
\describe{
  \item{X}{a \code{\link[Matrix]{sparseMatrix}} binary matrix of size 4475x965 with a one when one musicians played once in a given band.}
  \item{col_meta}{a data.frame with columns (bands) meta_data}
}
}
\source{
\href{http://deim.urv.cat/~alexandre.arenas/data/welcome.htm}{A. Arena Network datasets}
}
\usage{
data(Jazz_full)
}
\description{
List  Jazz musicians / Bands relationship
}
\examples{
data(Jazz_full)
}
\references{
P.Gleiser and L. Danon , Community Structure in jazz, Adv. Complex Syst.6, 565 (2003) (\href{https://arxiv.org/abs/cond-mat/0307434}{Arxiv})
}
\keyword{datasets}
