% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{K}
\alias{K}
\title{Generic method to extract the number of clusters found from an \code{\link{IclFit-class}} object}
\usage{
K(fit)
}
\arguments{
\item{fit}{an \code{IclFit} solution}
}
\value{
The number of clusters
}
\description{
This method take a \code{\link{IclFit-class}} object and return its ICL score.
}
