% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lca.R
\docType{class}
\name{LcaPrior-class}
\alias{LcaPrior-class}
\alias{LcaPrior}
\alias{Lca-class}
\alias{Lca}
\title{Latent Class Analysis Model Prior class}
\usage{
LcaPrior(beta = 1)

Lca(alpha = 1, beta = 1)
}
\description{
An S4 class to represent a Latent Class Analysis model
Such model can be used to cluster a data matrix \eqn{X} with the following generative model :  
\deqn{\pi&\sim \textrm{Dirichlet}(\alpha),}
\deqn{\forall k, \forall j, \quad \theta_{kj} &\sim \textrm{Dirichlet}_{d_j}(\beta),}
\deqn{Z_i&\sim \mathcal{M}_K(1,\pi),}
\deqn{\forall j=1, \ldots, p, \quad X_{ij}|Z_{ik}=1 &\sim \mathcal{M}_{d_j}(1, \theta_{kj}),}
}
\section{Functions}{
\itemize{
\item \code{LcaPrior}: LcaPrior class constructor

\item \code{Lca-class}: Lca class constructor

\item \code{Lca}: Lca class constructor
}}

\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter (default to 1)}

\item{\code{beta}}{Dirichlet over vocabulary prior parameter (default to 1)}
}}

\examples{
LcaPrior()
LcaPrior(beta = 0.5)
Lca()
Lca(beta = 0.5)
}
\seealso{
Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DcSbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{GmmPrior-class}},
\code{\link{MoMPrior-class}},
\code{\link{MoRPrior-class}},
\code{\link{MultSbmPrior-class}},
\code{\link{SbmPrior-class}}
}
\concept{DlvmModels}
