% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_post}
\alias{lm_post}
\title{lm_post}
\usage{
lm_post(X, y, regu, a0, b0)
}
\arguments{
\item{X}{data matrix of covariates Nxd}

\item{y}{target Nx1}

\item{regu}{prior precision parameter}

\item{a0}{prior parameter}

\item{b0}{prior parameter}
}
\description{
lm_post
}
