% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_post_add}
\alias{lm_post_add}
\title{lm_post_add}
\usage{
lm_post_add(current, X, y, regu, a0, b0)
}
\arguments{
\item{current}{gaussian linear model to update}

\item{X}{data matrix of covariates Ntxd}

\item{y}{target Ntx1}

\item{regu}{prior precision parameter}

\item{a0}{prior parameter}

\item{b0}{prior parameter}
}
\description{
lm_post_add
}
