% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_post_merge}
\alias{lm_post_merge}
\title{lm_post_merge}
\usage{
lm_post_merge(current_k, current_l, regu, a0, b0)
}
\arguments{
\item{current_k}{gaussian linear model to merge}

\item{current_l}{gaussian linear model to merge}

\item{regu}{prior precision parameter}

\item{a0}{prior parameter}

\item{b0}{prior parameter}
}
\description{
lm_post_merge
}
