% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_mat}
\alias{merge_mat}
\title{merge_mat}
\usage{
merge_mat(data, init_fit)
}
\arguments{
\item{data}{list with clustering data depnds on model type}

\item{init_fit}{initial fit object}
}
\value{
a cost merge matrix
}
\description{
merge_mat
}
