% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mixedmodels.R
\docType{class}
\name{MixedModels-class}
\alias{MixedModels-class}
\alias{MixedModels}
\title{Mixed Models classes}
\usage{
MixedModels(alpha = 1, models)
}
\description{
An S4 class to represent a mixed clustering models.
Such model can be used to cluster graph vertex, and model a square adjacency matrix \eqn{X} with the following generative model :
}
\section{Functions}{
\itemize{
\item \code{MixedModels}: MixedModels class constructor
}}

\section{Slots}{

\describe{
\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter (default to 10)}

\item{\code{p}}{Exponential prior parameter (default to NaN, in this case p will be estimated from data as the mean connection probability)}

\item{\code{type}}{define the type of networks (either "directed", "undirected" or "guess", default to "guess")}
}}

\examples{
MixedModels(models = list(continuous=Gmm(),discrete=Lca()))
}
\seealso{
Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DcSbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{GmmPrior-class}},
\code{\link{LcaPrior-class}},
\code{\link{MoMPrior-class}},
\code{\link{MoRPrior-class}},
\code{\link{MultSbmPrior-class}},
\code{\link{SbmPrior-class}}
}
\concept{DlvmModels}
