% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\docType{class}
\name{mm-class}
\alias{mm-class}
\title{Clustering with a mixture of multinomial model description class}
\description{
An S4 class to represent a multinomial model model, extend \code{\link{icl_model-class}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter (default to 1)}

\item{\code{beta}}{Dirchlet over vocabulary prior parameter (default to 1)}
}}

\examples{
new("mm")
new("mm",alpha=1,beta=1)
}
