% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_classes.R
\docType{class}
\name{models-classes}
\alias{models-classes}
\alias{icl_model}
\alias{icl_model-class}
\alias{sbm}
\alias{sbm-class}
\alias{mm}
\alias{mm-class}
\alias{dcsbm-class}
\title{Clustering models classes}
\description{
Clustering models classes

icl_model

An S4 class to represent an abstract clustering model
\itemize{
\item slots : \code{name,alpha}
}

sbm

An S4 class to represent a stochastick block model that extends \code{icl_model} class.
\itemize{
\item slots : \code{name,alpha,a0,b0}
}

sbm

An S4 class to represent a stochastick block model that extends \code{icl_model} class.
\itemize{
\item slots : \code{name,alpha,a0,b0}
}

mm

An S4 class to represent a mixture of multinomial also known has mixture of unigrams that extends \code{icl_model} class.
\itemize{
\item slots : \code{name,alpha,beta}
}
}
\section{Slots}{

\describe{
\item{\code{name}}{a character vector}

\item{\code{alpha}}{a numeric vector of length 1 which define the parameters of the dirichlet over the cluster proportions (default to 1)}

\item{\code{a0}}{a numeric vector of length 1 which define the parameters of the beta prior over the edges (default to 1)}

\item{\code{b0}}{a numeric vector of length 1 which define the parameters of the beta prior over the non-edges (default to 1)}

\item{\code{a0}}{a numeric vector of length 1 which define the parameters of the beta prior over the edges (default to 1)}

\item{\code{b0}}{a numeric vector of length 1 which define the parameters of the beta prior over the non-edges (default to 1)}

\item{\code{beta}}{a numeric vector of length 1 which define the parameters of the beta prior over the counts (default to 1)}
}}

\examples{
new("sbm")
new("sbm",a0=0.5,b0=0.5,alpha=1)
new("sbm")
new("sbm",a0=0.5,b0=0.5,alpha=1)
new("mm")
new("mm",alpha=1,beta=1)
}
