% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mom.R
\docType{class}
\name{MoMPrior-class}
\alias{MoMPrior-class}
\alias{MoMPrior}
\alias{MoM-class}
\alias{MoM}
\title{Mixture of Multinomial Model Prior description class}
\usage{
MoMPrior(beta = 1)

MoM(alpha = 1, beta = 1)
}
\description{
An S4 class to represent a Multinomial model model.
Such model can be used to cluster a data matrix \eqn{X} with the following generative model :
\deqn{ \pi \sim Dirichlet(\alpha)}
\deqn{ Z_i  \sim \mathcal{M}(1,\pi)}
\deqn{ \theta_{k} \sim Dirichlet(\beta)}
\deqn{ X_{i.}|Z_{ik}=1 \sim \mathcal{M}(L_i,\theta_{k})}
With \eqn{L_i=\sum_d=1^DX_{id}}. This class mainly store the prior parameters value (\eqn{\alpha,\beta}) of this generative model in the following slots:
}
\section{Functions}{
\itemize{
\item \code{MoMPrior}: MoMPrior class constructor

\item \code{MoM-class}: MoM class constructor

\item \code{MoM}: MoM class constructor
}}

\section{Slots}{

\describe{
\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter (default to 1)}

\item{\code{beta}}{Dirichlet over vocabulary prior parameter (default to 1)}
}}

\examples{
MoMPrior()
MoMPrior(beta = 0.5)
MoM()
MoM(beta = 0.5)
}
\seealso{
Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DcSbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{GmmPrior-class}},
\code{\link{LcaPrior-class}},
\code{\link{MixedModels-class}},
\code{\link{MoRPrior-class}},
\code{\link{MultSbmPrior-class}},
\code{\link{SbmPrior-class}}
}
\concept{DlvmModels}
