% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{class}
\name{Multistarts-class}
\alias{Multistarts-class}
\alias{Multistarts}
\title{Greedy algorithm with multiple start class}
\usage{
Multistarts(nb_start = 10)
}
\arguments{
\item{nb_start}{number of random starts (default to 10)}
}
\value{
a \code{Multistarts-class} object
}
\description{
An S4 class to represent a greedy algorithm  with multiple start (extends \code{\link{Alg-class}} class).
}
\section{Functions}{
\itemize{
\item \code{Multistarts()}: Multistarts algorithm class constructor

}}
\section{Slots}{

\describe{
\item{\code{nb_start}}{number of random starts (default to 10)}
}}

\examples{
Multistarts()
Multistarts(15)
}
