% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multsbm.R
\docType{class}
\name{multsbm-class}
\alias{multsbm-class}
\title{Clustering with a Multinomial Stochastic Block Model description class}
\description{
An S4 class to represent a Multinomial Stochastic Block Model, extend \code{\link{icl_model-class}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter}

\item{\code{beta}}{Dirichlet prior parameter over Multinomial links}
}}

