% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_multsbm.R
\docType{class}
\name{MultSbmPrior-class}
\alias{MultSbmPrior-class}
\alias{MultSbmPrior}
\alias{MultSbm-class}
\alias{MultSbm}
\title{Multinomial Stochastic Block Model Prior class}
\usage{
MultSbmPrior(beta = 1, type = "guess")

MultSbm(alpha = 1, beta = 1, type = "guess")
}
\description{
An S4 class to represent a Multinomial Stochastic Block Model. Such model can be used to cluster multilayer-graph vertex, and model a square adjacency cube \eqn{X} of size NxNxM with the following generative model :
\deqn{ \pi \sim Dirichlet(\alpha)}
\deqn{ Z_i  \sim \mathcal{M}(1,\pi)}
\deqn{ \theta_{kl} \sim Dirichlet(\beta)}
\deqn{ X_{ij.}|Z_{ik}Z_{jl}=1 \sim \mathcal{M}(L_{ij},\theta_{kl})}
With \eqn{L_{ij}=\sum_{m=1}^MX_{ijm}}. This class mainly store the prior parameters value \eqn{\alpha,\beta} of this generative model in the following slots:
}
\section{Functions}{
\itemize{
\item \code{MultSbmPrior}: MultSbmPrior class constructor

\item \code{MultSbm-class}: MultSbm class constructor

\item \code{MultSbm}: MultSbm class constructor
}}

\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter}

\item{\code{beta}}{Dirichlet prior parameter over Multinomial links}

\item{\code{type}}{define the type of networks (either "directed", "undirected" or "guess", default to "guess")}
}}

\examples{
MultSbmPrior()
MultSbmPrior(type = "undirected")
MultSbm()
MultSbm(type = "undirected")
}
\seealso{
\code{\link{MultSbmFit-class}}, \code{\link{MultSbmPath-class}}

Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DcSbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{GmmPrior-class}},
\code{\link{LcaPrior-class}},
\code{\link{MoMPrior-class}},
\code{\link{MoRPrior-class}},
\code{\link{SbmPrior-class}}
}
\concept{DlvmModels}
