% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mushroom}
\alias{mushroom}
\title{Mushroom data}
\format{
An R data.frame with a variable edibility used as label and 22
 categorical variables with no names. More detail on the UCI webpage
 describing the data.
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Mushroom}
}
\usage{
data(mushroom)
}
\description{
Categorical data from UCI Machine Learning Repository describing 8124
mushrooms with 22 phenotype variables. Each mushroom is classified as "edible"
or "poisonous" and the goal is to recover the mushroom class from its
phenotype.
}
\examples{
data(mushroom)
}
\keyword{datasets}
