% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmreg.R
\docType{class}
\name{mvmreg-class}
\alias{mvmreg-class}
\title{Multivariate mixture of regression model description class}
\description{
An S4 class to represent a multivariate mixture of regression model, extends \code{\link{icl_model-class}}.
#' The model follow \href{https://tminka.github.io/papers/minka-linear.pdf}{minka-linear}.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter (default to 1)}

\item{\code{beta}}{Prior parameter (inverse variance) default 0.01}

\item{\code{N0}}{Prior parameter (pseudo count) default to 10 ! should be > number of features}
}}

\examples{
new("mvmreg")
new("mvmreg",alpha=1,beta=0.1,N0=15)
}
