% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ndrangheta}
\alias{Ndrangheta}
\title{Ndrangheta mafia covert network dataset}
\format{

}
\source{
\href{https://sites.google.com/site/ucinetsoftware/datasets/covert-networks/ndranghetamafia2}{ucinetsoftware/datasets/covert-networks}
}
\usage{
data(Ndrangheta)
}
\description{
Network of co-attendance occurrence attendance of suspected members of the Ndrangheta criminal organization at summits (meetings whose purpose is to make important decisions and/or affiliations, but also to solve internal problems and to establish roles and powers) taking place between 2007 and 2009.
}
\examples{
data(Ndrangheta)
}
\references{
{Extended Stochastic Block Models with Application to Criminal Networks, Sirio Legramanti and Tommaso Rigon and Daniele Durante and David B. Dunson, 2021, 
(\href{https://arxiv.org/abs/2007.08569v2}{arXiv:2007.08569}).}
}
\keyword{datasets}
