% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NewGuinea}
\alias{NewGuinea}
\title{NewGuinea data}
\format{
A binary array of size (16,16,3) the first layer encodes enmity, the second, the friendship relations. The third, no relations between the two tribes.
}
\source{
\url{https://networks.skewed.de/net/new_guinea_tribes}
}
\usage{
data(NewGuinea)
}
\description{
\code{\link{NewGuinea}} a social network of 16 tribes, where two types of interactions were recorded, amounting to either friendship or enmity [read-cultures-1954].
}
\examples{
data(NewGuinea)
}
\references{
Kenneth E. Read, “Cultures of the Central Highlands, New Guinea”, Southwestern J. of Anthropology, 10(1):1-43 (1954). DOI: 10.1086/soutjanth.10.1.3629074
}
\keyword{datasets}
