% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nodelinklab}
\alias{nodelinklab}
\title{nodelinklab}
\usage{
nodelinklab(sol, labels, s = 0)
}
\arguments{
\item{sol}{\code{\link{mm_path-class}} object to be plot}

\item{labels}{a vector of cluster labels}

\item{s}{threshold for links}
}
\value{
a ggplot2 graph
}
\description{
nodelinklab
}
