% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclust_dcsbm.R
\name{plot,co_dcsbm_path,missing-method}
\alias{plot,co_dcsbm_path,missing-method}
\title{plot a \code{\link{co_dcsbm_path-class}}}
\usage{
\S4method{plot}{co_dcsbm_path,missing}(x, type = "blocks")
}
\arguments{
\item{x}{a \code{\link{co_dcsbm_path-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing connections between row and column clusters
\item \code{'nodelink'}: plot a nodelink diagram of the bipartite graph summarizing connections between row and column clusters
\item \code{'front'}: plot the extracted front ICL, log(alpha)
\item \code{'path'}: plot the evolution of ICL with respect to K
\item \code{'tree'}: plot the associated dendrograms one for the row clusters and one for the column clusters
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{co_dcsbm_path-class}}
}
