% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,dcsbm_fit,missing-method}
\alias{plot,dcsbm_fit,missing-method}
\title{Plot a clustering results}
\usage{
\S4method{plot}{dcsbm_fit,missing}(x, y, ..., type = "blocks")
}
\arguments{
\item{sol}{\code{\link{sbm_fit}} obejct to be ploted}
}
\value{
a ggplot2 graphics which summarize the results
}
\description{
Plot a clustering results
}
