% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dcsbm.R
\name{plot,DcSbmFit,missing-method}
\alias{plot,DcSbmFit,missing-method}
\title{Plot a \code{\link{DcSbmFit-class}} object}
\usage{
\S4method{plot}{DcSbmFit,missing}(x, type = "blocks")
}
\arguments{
\item{x}{a \code{\link{DcSbmFit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing connections between clusters
\item \code{'nodelink'}: plot a nodelink diagram of the graph summarizing connections between clusters
}}
}
\value{
a ggplot graphic
}
\description{
Plot a \code{\link{DcSbmFit-class}} object
}
\seealso{
\code{\link{plot,IclPath,missing-method}}
}
