% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diaggmm.R
\name{plot,diaggmm_fit,missing-method}
\alias{plot,diaggmm_fit,missing-method}
\title{plot a \code{\link{diaggmm_fit-class}} object}
\usage{
\S4method{plot}{diaggmm_fit,missing}(x, type = "marginals")
}
\arguments{
\item{x}{a \code{\link{diaggmm_fit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'marginals'}: plot the marginal densities 
\item \code{'violins'}: make a violin plot for each clusters and features 
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{diaggmm_fit-class}} object
}
