% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diaggmm.R
\name{plot,DiagGmmFit,missing-method}
\alias{plot,DiagGmmFit,missing-method}
\title{Plot a \code{\link{DiagGmmFit-class}} object}
\usage{
\S4method{plot}{DiagGmmFit,missing}(x, type = "marginals")
}
\arguments{
\item{x}{a \code{\link{DiagGmmFit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'marginals'}: plot the marginal densities
\item \code{'violins'}: make a violin plot for each clusters and features
}}
}
\value{
a ggplot graphic
}
\description{
Plot a \code{\link{DiagGmmFit-class}} object
}
