% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gmm.R
\name{plot,GmmFit,missing-method}
\alias{plot,GmmFit,missing-method}
\title{Plot a \code{\link{GmmFit-class}} object}
\usage{
\S4method{plot}{GmmFit,missing}(x, type = "marginals")
}
\arguments{
\item{x}{a \code{\link{GmmFit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'marginals'}: plot the marginal densities
\item \code{'violins'}: make a violin plot for each clusters and features
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
Plot a \code{\link{GmmFit-class}} object
}
