% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gmm.R
\name{plot,GmmPath,missing-method}
\alias{plot,GmmPath,missing-method}
\title{plot a \code{\link{GmmPath-class}} object}
\usage{
\S4method{plot}{GmmPath,missing}(x, type = "marginals")
}
\arguments{
\item{x}{a \code{\link{GmmPath-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'front'}: plot the extracted front ICL, log(alpha)
\item \code{'path'}: plot the evolution of ICL with respect to K
\item \code{'tree'}: plot the associated dendrogram
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{GmmPath-class}} object
}
