% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{plot,lca_fit,missing-method}
\alias{plot,lca_fit,missing-method}
\title{plot a \code{\link{lca_fit-class}} object}
\usage{
\S4method{plot}{lca_fit,missing}(x, type = "marginals")
}
\arguments{
\item{x}{a \code{\link{lca_fit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing connections between clusters
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{lca_fit-class}} object
}
