% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lca.R
\name{plot,LcaFit,missing-method}
\alias{plot,LcaFit,missing-method}
\title{Plot a \code{\link{LcaFit-class}} object}
\usage{
\S4method{plot}{LcaFit,missing}(x, type = "marginals")
}
\arguments{
\item{x}{a \code{\link{LcaFit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing connections between clusters
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
Plot a \code{\link{LcaFit-class}} object
}
