% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lca.R
\name{plot,LcaPath,missing-method}
\alias{plot,LcaPath,missing-method}
\title{plot a \code{\link{LcaPath-class}} object}
\usage{
\S4method{plot}{LcaPath,missing}(x, type = "marginals")
}
\arguments{
\item{x}{an \code{\link{LcaPath-class}} object}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'front'}: plot the extracted front in the plane ICL, log(alpha)
\item \code{'path'}: plot the evolution of ICL with respect to K
\item \code{'tree'}: plot the associated dendrogram
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{LcaPath-class}} object
}
