% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misssbm.R
\name{plot,misssbm_fit,missing-method}
\alias{plot,misssbm_fit,missing-method}
\title{plot a \code{\link{misssbm_fit-class}} object}
\usage{
\S4method{plot}{misssbm_fit,missing}(x, type = "blocks")
}
\arguments{
\item{x}{a \code{\link{misssbm_fit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing connections between clusters
\item \code{'nodelink'}: plot a nodelink diagram of the graph summarizing connections between clusters
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{misssbm_fit-class}} object
}
