% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,mm_fit,ANY-method}
\alias{plot,mm_fit,ANY-method}
\title{Plot a clustering results}
\usage{
\S4method{plot}{mm_fit,ANY}(x)
}
\arguments{
\item{sol}{\code{\link{mm_fit}} obejct to be ploted}
}
\value{
a ggplot2 graphics which summarize the results
}
\description{
Plot a clustering results
}
