% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{plot,mm_fit,missing-method}
\alias{plot,mm_fit,missing-method}
\title{plot a \code{\link{mm_fit-class}} object}
\usage{
\S4method{plot}{mm_fit,missing}(x, type = "blocks")
}
\arguments{
\item{x}{a \code{\link{mm_fit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing connections between clusters
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{mm_fit-class}} object
}
