% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\docType{methods}
\name{plot,mm_path,missing-method}
\alias{plot,mm_path,missing-method}
\title{plot a \code{\link{mm_path-class}} object}
\usage{
\S4method{plot}{mm_path,missing}(x, type = "blocks")
}
\arguments{
\item{x}{an \code{\link{mm_path-class}} object}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing 
\item \code{'nodelink'}: plot a nodelink diagram 
\item \code{'front'}: plot the extracted front in the plane ICL, log(alpha)
\item \code{'path'}: plot the veolution of ICL with repsect to K
\item \code{'tree'}: plot the associated dendogram
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{mm_path-class}} object
}
