% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mom.R
\name{plot,MoMFit,missing-method}
\alias{plot,MoMFit,missing-method}
\title{Plot a \code{\link{MoMFit-class}} object}
\usage{
\S4method{plot}{MoMFit,missing}(x, type = "blocks")
}
\arguments{
\item{x}{a \code{\link{MoMFit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing connections between clusters
}}
}
\value{
a ggplot graphic
}
\description{
Plot a \code{\link{MoMFit-class}} object
}
