% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mom.R
\name{plot,MoMPath,missing-method}
\alias{plot,MoMPath,missing-method}
\title{plot a \code{\link{MoMPath-class}} object}
\usage{
\S4method{plot}{MoMPath,missing}(x, type = "blocks")
}
\arguments{
\item{x}{an \code{\link{MoMPath-class}} object}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing
\item \code{'nodelink'}: plot a nodelink diagram
\item \code{'front'}: plot the extracted front in the plane ICL, log(alpha)
\item \code{'path'}: plot the evolution of ICL with respect to K
\item \code{'tree'}: plot the associated dendrogram
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{MoMPath-class}} object
}
