% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_multsbm.R
\name{plot,MultSbmFit,missing-method}
\alias{plot,MultSbmFit,missing-method}
\title{Plot a \code{\link{MultSbmFit-class}} object}
\usage{
\S4method{plot}{MultSbmFit,missing}(x, type = "blocks")
}
\arguments{
\item{x}{a \code{\link{MultSbmFit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing connections between clusters
\item \code{'nodelink'}: plot a nodelink diagram of the graph summarizing connections between clusters
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
Plot a \code{\link{MultSbmFit-class}} object
}
