% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmreg.R
\docType{methods}
\name{plot,mvmreg_path,missing-method}
\alias{plot,mvmreg_path,missing-method}
\title{plot a \code{\link{mvmreg_path-class}} object}
\usage{
\S4method{plot}{mvmreg_path,missing}(x, type = "tree")
}
\arguments{
\item{x}{a \code{\link{mvmreg_path-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'front'}: plot the extracted front ICL, log(alpha)
\item \code{'path'}: plot the veolution of ICL with repsect to K
\item \code{'tree'}: plot the associated dendogram
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{mvmreg_path-class}} object
}
