% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{plot,sbm_path,missing-method}
\alias{plot,sbm_path,missing-method}
\title{plot a \code{\link{sbm_path-class}} object}
\usage{
\S4method{plot}{sbm_path,missing}(x, type = "blocks")
}
\arguments{
\item{x}{an \code{\link{sbm_path-class}} object}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing connections between clusters
\item \code{'nodelink'}: plot a nodelink diagram of the bipartite graph summarizing connections between clusters
\item \code{'front'}: plot the extracted front ICL, log(alpha)
\item \code{'path'}: plot the evolution of ICL with respect to K
\item \code{'tree'}: plot the associated dendrogram
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
plot a \code{\link{sbm_path-class}} object
}
