% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sbm.R
\name{plot,SbmFit,missing-method}
\alias{plot,SbmFit,missing-method}
\title{Plot a \code{\link{SbmFit-class}} object}
\usage{
\S4method{plot}{SbmFit,missing}(x, type = "blocks")
}
\arguments{
\item{x}{a \code{\link{SbmFit-class}}}

\item{type}{a string which specify plot type:
\itemize{
\item \code{'blocks'}: plot a block matrix with summarizing connections between clusters
\item \code{'nodelink'}: plot a nodelink diagram of the graph summarizing connections between clusters
}}
}
\value{
a \code{\link{ggplot2}} graphic
}
\description{
Plot a \code{\link{SbmFit-class}} object
}
\seealso{
\code{\link{plot,IclPath,missing-method}}
}
