% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsbm.R, R/mm.R, R/mreg.R, R/mvmreg.R,
%   R/plot.R, R/sbm.R
\docType{methods}
\name{plot,dcsbm_fit,missing-method}
\alias{plot,dcsbm_fit,missing-method}
\alias{plot,dcsbm_path,missing-method}
\alias{plot,mm_fit,missing-method}
\alias{plot,mm_path,missing-method}
\alias{plot,mreg_fit,missing-method}
\alias{plot,mreg_path,missing-method}
\alias{plot,mvmreg_path,missing-method}
\alias{plot}
\alias{plot,sbm_fit,missing-method}
\alias{plot,sbm_path,missing-method}
\title{Plot a clustering results}
\usage{
\S4method{plot}{dcsbm_fit,missing}(x, type = "blocks")

\S4method{plot}{dcsbm_path,missing}(x, type = "blocks")

\S4method{plot}{mm_fit,missing}(x, type = "blocks")

\S4method{plot}{mm_path,missing}(x, type = "blocks")

\S4method{plot}{mreg_fit,missing}(x, type = "blocks")

\S4method{plot}{mreg_path,missing}(x, type = "blocks")

\S4method{plot}{mvmreg_path,missing}(x, type = "blocks")

\S4method{plot}{sbm_fit,missing}(x, type = "blocks")

\S4method{plot}{sbm_path,missing}(x, type = "blocks")
}
\arguments{
\item{x}{\code{\link{icl_fit-class}} object to be ploted}

\item{type}{type of desired graphics : tree,pathy, blocks, nodelink, front}
}
\value{
a ggplot2 object to visualize the results
}
\description{
Main methods to explore clusterings results visualy.
}
