% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{post_probs}
\alias{post_probs}
\title{post_probs}
\usage{
post_probs(model, xp, clt)
}
\arguments{
\item{model}{icl_model}

\item{xp}{sparseMatrix}

\item{clt}{cluster labels {0,...,K-1}}
}
\description{
post_probs
}
