% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{methods}
\name{postprocess,icl_path-method}
\alias{postprocess,icl_path-method}
\title{postprocess}
\usage{
\S4method{postprocess}{icl_path}(path, data = NULL)
}
\arguments{
\item{path}{an icl_path object}
}
\description{
postprocess
}
