% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{methods}
\name{preprocess,icl_model-method}
\alias{preprocess,icl_model-method}
\title{preprocess}
\usage{
\S4method{preprocess}{icl_model}(model, data)
}
\arguments{
\item{model}{an icl_model}

\item{X}{input data to prepare}
}
\description{
preprocess
}
