% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{print,icl_path-method}
\alias{print,icl_path-method}
\title{print an icl_path object}
\usage{
\S4method{print}{icl_path}(x)
}
\arguments{
\item{x}{\code{\link{icl_path-class}} object to print}
}
\value{
None (invisible NULL). No return value, called for side effects.
}
\description{
Print an \code{\link{icl_path-class}} object, model type and number of found clusters are provided.
}
