% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_ploting.R
\name{print,IclPath-method}
\alias{print,IclPath-method}
\title{print an IclPath object}
\usage{
\S4method{print}{IclPath}(x)
}
\arguments{
\item{x}{\code{\link{IclPath-class}} object to print}
}
\value{
None (invisible NULL). No return value, called for side effects.
}
\description{
Print an \code{\link{IclPath-class}} object, model type and number of found clusters are provided.
}
