% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{prior,IclFit-method}
\alias{prior,IclFit-method}
\title{Method to extract the prior used to fit  \code{\link{IclFit-class}} object}
\usage{
\S4method{prior}{IclFit}(fit)
}
\arguments{
\item{fit}{an \code{IclFit} solution}
}
\value{
The prior an S4 object of the relevant class with the prior parameters
}
\description{
This method take a \code{\link{IclFit-class}} object and return the prior used.
}
\seealso{
\code{\link{Sbm-class}},\code{\link{Gmm-class}},...
}
