% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{reorder_sbm}
\alias{reorder_sbm}
\title{#' @describeIn fit 
#' @export
setMethod(f = "fit", 
          signature = signature("dgCMatrix", "numeric"),
          definition = function(x,K,verbose=FALSE){
            # only a sparseMatrix and a number check dim to choose between graph models and mm
            if(dim(x)[1]==dim(x)[2]){
              fit(x,K,new("dcsbm"),new("hybrid"),verbose=verbose)  
            }else{
              fit(x,K,new("mm"),new("hybrid"),verbose=verbose)
            }
          });}
\usage{
reorder_sbm(obs_stats, or)
}
\description{
#' @describeIn fit 
#' @export
setMethod(f = "fit", 
          signature = signature("dgCMatrix", "missing", "numeric","icl_model", "missing"),
          definition = function(x,K,model,verbose=FALSE){
            fit(x,K,model,new("hybrid"),verbose=verbose)
          });
}
