% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{rmm}
\alias{rmm}
\title{Generate graph adjacency matrix using a SBM}
\usage{
rmm(N, pi, mu, lambda)
}
\arguments{
\item{N}{A numeric value the size of the graph to generate}

\item{pi}{A numeric vector of length K with clusters proportions. Must sum up to 1.}

\item{mu}{A numeric matrix of dim k x K with the connectivity pattern to generate. elements in [0,1].}

\item{lambda}{A numeric value which specify the expectation}
}
\value{
A list with fields:
\itemize{
\item x:
\item K:
\item N:
\item cl:
\item pi:
\item mu: 
\item lambda:
}
}
\description{
\code{rmm} returns a count matrix and the cluster labels generated randomly unsig a Mixture of Multinomial model.
}
\details{
It take the sample size, cluster proportions and emission matrix, and  as input and sample a graph accordingly together with the clusters labels.
}
