% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{rsbm}
\alias{rsbm}
\title{Generate graph adjacency matrix using a SBM}
\usage{
rsbm(N, pi, mu)
}
\arguments{
\item{N}{A numeric value the size of the graph to generate}

\item{pi}{A numeric vector of length K with clusters proportions. Must sum up to 1.}

\item{mu}{A numeric matrix of dim k x K with the connectivity pattern to generate. elements in [0,1].}
}
\value{
A list with fields:
\itemize{
\item x:
\item K:
\item N:
\item cl:
\item pi:
\item mu: 
}
}
\description{
\code{rsbm} returns the adjacency matrix and the cluster labels generated randomly unsing a Stochastick Block Model.
}
\details{
Thisfunction take graph size, cluster proportions and connectivity matrix as input and sample a graph accordingly together with the clusters labels.
}
\examples{
simu = rsbm(100,rep(1/5,5),diag(rep(0.1,5))+0.001)
x  = simu$x
xl = simu$cl
}
