% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\docType{class}
\name{sbm_fit-class}
\alias{sbm_fit-class}
\title{sbm_fit

An S4 class to represent a fit of a stochastick block model that extend \code{icl_fit}.}
\description{
sbm_fit

An S4 class to represent a fit of a stochastick block model that extend \code{icl_fit}.
}
\section{Slots}{

\describe{
\item{\code{model}}{an \code{\link{sbm}} to store the model fitted}
}}

