% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\docType{class}
\name{sbm_path-class}
\alias{sbm_path-class}
\title{sbm_path

An S4 class to represent a hierachical path of solutions for a SBM model that extend \code{sbm_fit-class} and \code{icl_path-class}.}
\description{
sbm_path

An S4 class to represent a hierachical path of solutions for a SBM model that extend \code{sbm_fit-class} and \code{icl_path-class}.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{K}}{number of extracted cluster}

\item{\code{icl}}{value of the icl criterion}

\item{\code{cl}}{vector of cluster label}

\item{\code{obs_stats}}{list of model observed statistics (counts: number of nodes per clusters, x_counts: matrix with number of edges between each cluster)}

\item{\code{model}}{\code{\link{sbm}} model used with it's prior parameters}

\item{\code{path}}{list of clustering solution build using hierachical greedy merge}

\item{\code{tree}}{merge tree}

\item{\code{ggtree}}{data.frame with merge tree meta information}

\item{\code{logalpha}}{value of log(alpha) were alpha is the dirichlet prior parameter on cluster proportions}

\item{\code{train_hist}}{data.frame with training history information}

\item{\code{move_mat}}{sparse.matrix with eventually the constraints used for merge and swap moves}
}}

