% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sbm.R
\docType{class}
\name{SbmPrior-class}
\alias{SbmPrior-class}
\alias{SbmPrior}
\alias{Sbm-class}
\alias{Sbm}
\title{Stochastic Block Model Prior class}
\usage{
SbmPrior(a0 = 1, b0 = 1, type = "guess")

Sbm(alpha = 1, a0 = 1, b0 = 1, type = "guess")
}
\description{
An S4 class to represent a Stochastic Block Model.
Such model can be used to cluster graph vertex, and model a square adjacency matrix \eqn{X} with the following generative model :
\deqn{ \pi \sim Dirichlet(\alpha)}
\deqn{ Z_i  \sim \mathcal{M}(1,\pi)}
\deqn{ \theta_{kl} \sim Beta(a_0,b_0)}
\deqn{ X_{ij}|Z_{ik}Z_{jl}=1 \sim \mathcal{B}(\theta_{kl})}
This class mainly store the prior parameters value \eqn{\alpha,a_0,b_0} of this generative model in the following slots:
}
\section{Functions}{
\itemize{
\item \code{SbmPrior}: SbmPrior class constructor

\item \code{Sbm-class}: Sbm class constructor

\item \code{Sbm}: Sbm class constructor
}}

\section{Slots}{

\describe{
\item{\code{a0}}{Beta prior parameter over links (default to 1)}

\item{\code{b0}}{Beta prior parameter over no-links (default to 1)}

\item{\code{type}}{define the type of networks (either "directed", "undirected" or "guess", default to "guess"), for undirected graphs the adjacency matrix is supposed to be symmetric.}
}}

\examples{
Sbm()
sbm <- rsbm(100, c(0.5, 0.5), diag(2) * 0.1 + 0.01)
sol <- greed(sbm$x, model = Sbm(a0 = 0.5, b0 = 0.5, alpha = 0.5))
SbmPrior()
SbmPrior(type = "undirected")
Sbm()
Sbm(type = "undirected")
}
\references{
Nowicki, Krzysztof and Tom A B Snijders (2001). “Estimation and prediction for stochastic block structures”. In:Journal of the American statistical association 96.455, pp. 1077–1087
}
\seealso{
\code{\link{SbmFit-class}},\code{\link{SbmPath-class}}

\code{\link{greed}}

Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DcSbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{GmmPrior-class}},
\code{\link{LcaPrior-class}},
\code{\link{MixedModels-class}},
\code{\link{MoMPrior-class}},
\code{\link{MoRPrior-class}},
\code{\link{MultSbmPrior-class}}
}
\concept{DlvmModels}
