% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{class}
\name{seed-class}
\alias{seed-class}
\title{greedy algorithm with seeded initialisation}
\description{
An S4 class to represent a greedy algorithm with initialization from spectral clustering and or k-means (extends \code{\link{alg-class}} class ).
}
