% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_ploting.R
\name{show,IclFit-method}
\alias{show,IclFit-method}
\title{Show an IclPath object}
\usage{
\S4method{show}{IclFit}(object)
}
\arguments{
\item{object}{\code{\link{IclPath-class}} object to print}
}
\value{
None (invisible NULL). No return value, called for side effects.
}
\description{
Print an \code{\link{IclPath-class}} object, model type and number of found clusters are provided.
}
