% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{show,icl_path-method}
\alias{show,icl_path-method}
\title{show
show an icl_path object}
\usage{
\S4method{show}{icl_path}(object)
}
\arguments{
\item{object}{\code{\link{icl_path-class}} object to print}
}
\description{
show
show an icl_path object
}
