% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{spectral}
\alias{spectral}
\title{spectral
Regularized spectral clustering nips paper 2013}
\usage{
spectral(X, K)
}
\arguments{
\item{K}{Desired number of cluster}

\item{x}{An adjacency matrix in sparse format}
}
\description{
spectral
Regularized spectral clustering nips paper 2013
}
