% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_misc.R
\name{to_multinomial}
\alias{to_multinomial}
\title{Convert a binary adjacency matrix with missing value to a cube}
\usage{
to_multinomial(X)
}
\arguments{
\item{X}{A binary adjacency matrix with NA}
}
\value{
a cube
}
\description{
Convert a binary adjacency matrix with missing value to a cube
}
