% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valcovid}
\alias{valcovid}
\title{Smart card validation dataset}
\format{
An object of class \code{list} with two fields; 
 \describe{
  \item{data_clean}{row data}
  \item{X}{matrix}
}
}
\usage{
data(valcovid)
}
\description{
Smart card validation dataset
}
\examples{
data(valcovid)
}
\keyword{datasets}
