% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Xvlegislature}
\alias{Xvlegislature}
\title{French Parliament votes dataset}
\format{
An object of class \code{list} with three fields; 
 \describe{
  \item{X}{matrix of deputy votes a \code{\link[Matrix]{sparseMatrix}} of size 593x1839}
  \item{rowmeta}{a data frame with deputy meta-data}
  \item{colmeta}{a data frame with votes meta data}
}
}
\usage{
data(Xvlegislature)
}
\description{
French Parliament votes dataset
}
\examples{
data(Xvlegislature)
}
\keyword{datasets}
