% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Youngpeoplesurvey}
\alias{Youngpeoplesurvey}
\title{Young People survey data}
\format{
An R data.frame with columns containing each of the 150 original
  variables of the study.
}
\source{
\url{https://www.kaggle.com/miroslavsabo/young-people-survey}
}
\usage{
data(Youngpeoplesurvey)
}
\description{
Young people survey data from Miroslav Sabo and available on the Kaggle
platform. This is an authentic example of questionnaire data where Slovakian
young people (15-30 years old) were asked musical preferences according to
different genres (rock, hip-hop, classical, etc.).
}
\examples{
data(Youngpeoplesurvey)
}
\keyword{datasets}
