% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclust_dcsbm.R
\docType{class}
\name{co_dcsbm-class}
\alias{co_dcsbm-class}
\title{Co-clustering with a degree correted stochastick block model class}
\description{
An S4 class to represent a degree corrected stochastick block model for co_clustering, extends \code{\link{icl_model-class}} class.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{dirichlet parameters for the prior over clusters proportions (default to 1)}
}}

\examples{
new("co_dcsbm")
}
