% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{cut,icl_path-method}
\alias{cut,icl_path-method}
\title{Method to cut a path solution to a desired number of cluster}
\usage{
\S4method{cut}{icl_path}(x, K)
}
\arguments{
\item{x}{A an \code{icl_path} solution}

\item{K}{Desired number of cluster}
}
\value{
an \code{\link{icl_path-class}} object with the desired number of cluster
}
\description{
This method take a \code{\link{icl_path-class}} object and an integer K and return the solution from the path with K clusters
}
