% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mixedmodels.R
\name{extractSubModel}
\alias{extractSubModel}
\title{extract a part of a \code{\link{MixedModelsPath-class}} object}
\usage{
extractSubModel(sol, sub_model_name)
}
\arguments{
\item{sol}{an \code{\link{MixedModelsPath-class}} object}

\item{sub_model_name}{a string which specify the part of the model to extract}
}
\value{
a \code{\link{IclFit-class}} object of the relevant class
}
\description{
extract a part of a \code{\link{MixedModelsPath-class}} object
}
