% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fashion}
\alias{fashion}
\title{Fashion mnist dataset}
\format{
An object of class \code{list} with two fields; 
 \describe{
  \item{X}{pixels intensities values data.frame of size 1 000x784}
  \item{labels}{vector of labels with the following encoding (0 = T-shirt/top,
  1 = Trouser,
  2 = Pullover,
  3 = Dress,
  4 = Coat,
  5 = Sandal,
  6 = Shirt,
  7 = Sneaker,
  8 = Bag
  9 = Ankle boot)}
}
}
\source{
\href{https://github.com/zalandoresearch/fashion-mnist}{https://github.com/zalandoresearch/fashion-mnist}
}
\usage{
data(fashion)
}
\description{
Zalando fashionmnist dataset, sample of 1 000 Zalando's article images from the test set.
}
\examples{
data(fashion)
}
\references{
Fashion-MNIST: a Novel Image Dataset for Benchmarking Machine Learning Algorithms. Han Xiao, Kashif Rasul, Roland Vollgraf (2017)  
(\href{https://arxiv.org/abs/1708.07747}{arXiv:1708.07747}).
}
\keyword{datasets}
