% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fifa}
\alias{Fifa}
\title{Fifa data}
\format{
An R data.frame with columns containing each of the descriptive
  statistics of a player.
}
\source{
\url{https://www.kaggle.com/stefanoleone992/fifa-20-complete-player-dataset?select=players_20.csv}
}
\usage{
data(Fifa)
}
\description{
A random sample of 6000 players from the FIFA videogame with various statistics on all player ranging
from position, cost in the game, capacity in offense/defense, speed, etc. 
Two columns pos_x, pos_y with average player possible positions (in opta coordiantes)
were derived from the raw data.  was also u.
}
\examples{
data(Fifa)
}
\keyword{datasets}
