% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,dgCMatrix,numeric,missing,missing-method}
\alias{fit,dgCMatrix,numeric,icl_model,missing-method}
\alias{fit,dgCMatrix,numeric,icl_model,genetic-method}
\alias{fit,dgCMatrix,numeric,icl_model,greed-method}
\alias{fit,dgCMatrix,numeric,icl_model,km-method}
\title{Fit a clustering model}
\usage{
fit(x, K, model, alg, ...)

\S4method{fit}{dgCMatrix,numeric,missing,missing}(x, K, verbose = FALSE)

\S4method{fit}{dgCMatrix,numeric,icl_model,missing}(x, K, model,
  verbose = FALSE)

\S4method{fit}{dgCMatrix,numeric,icl_model,genetic}(x, K, model, alg,
  verbose = FALSE)

\S4method{fit}{dgCMatrix,numeric,icl_model,greed}(x, K, model, alg,
  verbose = FALSE)

\S4method{fit}{dgCMatrix,numeric,icl_model,km}(x, K, model, alg,
  verbose = FALSE)
}
\arguments{
\item{x}{A sparse Matrix as a \code{dgCMatrix}}

\item{K}{An initial guess of the maximal number of cluster}

\item{model}{An \code{\link{IclModel-class}} such as \code{\link{sbm-class}}, \code{\link{dcsbm-class}}, ...}

\item{alg}{An optimization algorithm such as \code{\link{greed-class}}, \code{\link{genetic-class}} or \code{\link{km-class}}}
}
\description{
Fit a clustering model
}
\section{Methods (by class)}{
\itemize{
\item \code{x = dgCMatrix,K = numeric,model = missing,alg = missing}: 

\item \code{x = dgCMatrix,K = numeric,model = icl_model,alg = missing}: 

\item \code{x = dgCMatrix,K = numeric,model = icl_model,alg = genetic}: 

\item \code{x = dgCMatrix,K = numeric,model = icl_model,alg = greed}: 

\item \code{x = dgCMatrix,K = numeric,model = icl_model,alg = km}: 
}}

