% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{fit_cond}
\alias{fit_cond}
\title{Fit a clustering model}
\usage{
fit_cond(model, alg, ...)
}
\arguments{
\item{model}{An \code{\link{IclModel-class}} such as \code{\link{sbm-class}}, \code{\link{dcsbm-class}}, ...}

\item{alg}{An optimization algorithm such as \code{\link{greed-class}}, \code{\link{genetic-class}} or \code{\link{km-class}}}

\item{x}{A sparse Matrix as a \code{dgCMatrix}}

\item{K}{An initial guess of the maximal number of cluster}
}
\description{
Fit a clustering model
}
