% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_greed}
\alias{fit_greed}
\title{fit_greed}
\usage{
fit_greed(model, data, clt, type = "both", nb_max_pass = 50L,
  verbose = FALSE)
}
\arguments{
\item{model}{icl_model}

\item{data}{list with clustering data (fileds depend on model type)}

\item{clt}{cluster labels {0,...,K-1}}

\item{type}{merge, swap, none, or both (default)}

\item{nb_max_pass}{maximum number of pass for greedy swap}

\item{verbose}{boolean for verbose mode default to false}
}
\value{
a model_fit object
}
\description{
fit_greed
}
