% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_greed}
\alias{fit_greed}
\title{fit_greed}
\usage{
fit_greed(model, xp, Ki)
}
\arguments{
\item{model}{icl_model}

\item{xp}{sparseMatrix}

\item{Ki}{initia guess for K}
}
\description{
fit_greed
}
