% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_greed}
\alias{fit_greed}
\title{fit_greed_init}
\usage{
fit_greed(model, data, clt, type = "both", nb_max_pass = 50L,
  verbose = FALSE)
}
\arguments{
\item{model}{icl_model}

\item{clt}{cluster labels {0,...,K-1}}

\item{type}{: merge, swap, none, or both (default)}

\item{nb_max_pass}{: maximum number of pass for greedy swap}

\item{xp}{sparseMatrix}
}
\description{
fit_greed_init
}
