% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_greed_init}
\alias{fit_greed_init}
\title{fit_greed_init_type}
\usage{
fit_greed_init(model, xp, clt, type = "both", nb_max_pass = 50L,
  verbose = FALSE)
}
\arguments{
\item{model}{icl_model}

\item{xp}{sparseMatrix}

\item{clt}{cluster labels {0,...,K-1}}

\item{type}{: merge, swap or both (default)}

\item{nb_max_pass}{: maximum number of pass for greedy swap}
}
\description{
fit_greed_init_type
}
