% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Football}
\alias{Football}
\title{American College football network dataset}
\format{
An object of class \code{list} with two fields;
 \describe{
  \item{X}{network adjacency matrix as a \code{\link[Matrix]{sparseMatrix}} of size 115x115}
  \item{label}{vector of teams conferences of size 115 with the following encoding (0 = Atlantic Coast,
  1 = Big East,
  2 = Big Ten,
  3 = Big Twelve,
  4 = Conference USA,
  5 = Independents,
  6 = Mid-American,
  7 = Mountain West,
  8 = Pacific Ten,
  9 = Southeastern,
  10 = Sun Belt,
  11 = Western Athletic)}
}
}
\usage{
data(Football)
}
\description{
Network of American football games between Division IA colleges during regular season Fall 2000.
}
\examples{
data(Football)
}
\references{
M. Girvan and M. E. J. Newman, Community structure in social and biological networks, Proc. Natl. Acad. Sci. USA 99, 7821-7826 (2002)
}
\keyword{datasets}
