% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\docType{class}
\name{gmm-class}
\alias{gmm-class}
\title{Clustering with a gaussian mixture model description class}
\description{
An S4 class to represent a multivariate mixture of regression model, extend \code{\link{icl_model-class}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter (default to 1)}

\item{\code{tau}}{Prior parameter (inverse variance) default 0.01}

\item{\code{N0}}{Prior parameter (pseudo count) defulat to 10 ! should be > number of features}

\item{\code{epsilon}}{Prior parameter covartiance matrix prior}

\item{\code{mu}}{mean prior}
}}

\examples{
new("gmm")
new("gmm",alpha=1,tau=0.1,N0=15)
}
