% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R, R/greed.R
\docType{package}
\name{greed}
\alias{greed}
\alias{greed-package}
\title{model based hierachical clustering}
\usage{
greed(X, K = 20, model = find_model(X), alg = methods::new("hybrid"),
  verbose = FALSE)
}
\arguments{
\item{X}{data to cluster}

\item{K}{Desired number of cluster}

\item{alg}{an optimisation algorithm of class \code{\link{hybrid-class}} (default), \code{\link{multistarts-class}}, \code{\link{seed-class}} or \code{\link{genetic-class}}}

\item{verbose}{boolean for verbose mode}
}
\value{
an \code{\link{icl_path-class}} object
}
\description{
Performs  a

Greed enable the clustering of networks and counts data such as document/term matrix with different model. 
Model selection and clustering is performed in combination by optimizing the Integrated Classification Likelihood 
(which is equivalent to minimizing the description length). 
Their are four models availables : 
\itemize{
\item Stochastic Block Model (directed)
\item Degree corected Stochastic Block Model (directed)
\item Mixture of Multinomials
\item Multivariate mixture of poissons
}
The optimization is performed thanks to a combination of greedy local search and a genetic algorithm. 
The main entry point is the \code{\link{greed}} function to perfom the clustering.
}
