% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R, R/greed.R
\docType{package}
\name{greed}
\alias{greed}
\title{model based hierarchical clustering with icl}
\usage{
greed(
  X,
  K = 20,
  model = find_model(X),
  alg = methods::new("hybrid"),
  verbose = FALSE
)
}
\arguments{
\item{X}{data to cluster either a matrix or a \code{\link{dgCMatrix-class}}}

\item{K}{initial number of cluster}

\item{model}{a generative model to fit \code{\link{sbm-class}}, \code{\link{dcsbm-class}}, \code{\link{co_dcsbm-class}}, \code{\link{mm-class}} or \code{\link{mvmreg-class}}}

\item{alg}{an optimization algorithm of class \code{\link{hybrid-class}} (default), \code{\link{multistarts-class}}, \code{\link{seed-class}} or \code{\link{genetic-class}}}

\item{verbose}{Boolean for verbose mode}
}
\value{
an \code{\link{icl_path-class}} object
}
\description{


Greed enable the clustering of networks and counts data such as document/term matrix with different model. 
Model selection and clustering is performed in combination by optimizing the Integrated Classification Likelihood 
(which is equivalent to minimizing the description length). 
Their following generative models are available : 
\itemize{
\item Stochastic Block Model (directed)
\item Degree corrected Stochastic Block Model (directed)
\item Mixture of Multinomial
\item Latent Block Model (directed)
\item Degree corrected Latent Block Model (directed)
\item Gaussian Mixture Model (experimental)
\item Mixture of Multivariate Regressions (experimental)
}
The optimization is performed thanks to a combination of greedy local search and a genetic algorithm. 
The main entry point is the \code{\link{greed}} function to perform the clustering.
}
