% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R, R/greed.R
\docType{package}
\name{greed}
\alias{greed}
\title{Model based hierarchical clustering}
\usage{
greed(X, model = find_model(X), K = 20, alg = Hybrid(), verbose = FALSE)
}
\arguments{
\item{X}{data to cluster either a data.frame, a matrix, an array or a \code{\link{dgCMatrix-class}}}

\item{model}{a generative model to fit \code{\link{DlvmPrior-class}}}

\item{K}{initial number of cluster}

\item{alg}{an optimization algorithm of class \code{\link{Alg-class}} such as \code{\link{Hybrid-class}} (default), \code{\link{Multistarts-class}}, \code{\link{Seed-class}} or \code{\link{Genetic-class}}}

\item{verbose}{Boolean for verbose mode}
}
\value{
an \code{\link{IclPath-class}} object
}
\description{


Greed enables the clustering of networks and count data such as document-term
matrix with different models. Model selection and clustering are performed in
combination by optimizing the Integrated Classification Likelihood.
Optimization is performed thanks to a combination of greedy local search and
a genetic algorithm. The main entry point is the \code{\link{greed}} function
to perform the clustering, which is documented below. The package also
provides sampling functions for all the implemented DLVMs
}
\seealso{
Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DcSbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{DlvmPrior-class}},
\code{\link{GmmPrior-class}},
\code{\link{LcaPrior-class}},
\code{\link{MixedModels-class}},
\code{\link{MoMPrior-class}},
\code{\link{MoRPrior-class}},
\code{\link{MultSbmPrior-class}},
\code{\link{SbmPrior-class}}
}
\concept{DlvmModels}
