% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{greed_cond}
\alias{greed_cond}
\title{greed_cond}
\usage{
greed_cond(X, y, K = 20, model = find_model_cond(X, y),
  alg = methods::new("hybrid"), verbose = FALSE)
}
\arguments{
\item{X}{covariable data}

\item{y}{target variable}

\item{K}{Desired number of cluster}

\item{model}{an mreg model}

\item{alg}{an optimisation algorithm of class hybrid, multistarts, seed or genetic}

\item{verbose}{boolean for verbose mode}
}
\value{
an icl_path object
}
\description{
greed_cond
}
