% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{class}
\name{Hybrid-class}
\alias{Hybrid-class}
\alias{Hybrid}
\title{Hybrid optimization algorithm}
\usage{
Hybrid(pop_size = 20, nb_max_gen = 10, prob_mutation = 0.25, Kmax = 100)
}
\arguments{
\item{pop_size}{size of the solutions populations (default to 20)}

\item{nb_max_gen}{maximal number of generation to produce (default to 10)}

\item{prob_mutation}{mutation probability (default to 0.25)}

\item{Kmax}{maximum number of clusters (default to 100)}
}
\value{
a \code{Hybrid-class} object
}
\description{
An S4 class to represent an hybrid genetic/greedy algorithm (extends \code{\link{Alg-class}} class).
}
\section{Functions}{
\itemize{
\item \code{Hybrid}: Hybrid algorithm class constructor
}}

\section{Slots}{

\describe{
\item{\code{pop_size}}{size of the solutions populations (default to 20)}

\item{\code{nb_max_gen}}{maximal number of generation to produce (default to 10)}

\item{\code{prob_mutation}}{mutation probability (default to 0.25)}

\item{\code{Kmax}}{maximum number of clusters (default to 100)}
}}

\examples{
Hybrid()
Hybrid(pop_size = 100)
}
