% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{K}
\alias{K}
\alias{K,IclFit-method}
\title{Generic method to get the number of clusters from an \code{\link{IclFit-class}} object}
\usage{
K(fit)

\S4method{K}{IclFit}(fit)
}
\arguments{
\item{fit}{an \code{IclFit} solution}
}
\value{
The number of clusters
}
\description{
This method take a \code{\link{IclFit-class}} object and return its ICL score.
}
\section{Methods (by class)}{
\itemize{
\item \code{IclFit}: IclFit method
}}

