% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mixedmodels.R
\docType{class}
\name{MixedModels-class}
\alias{MixedModels-class}
\alias{MixedModels}
\title{Mixed Models classes}
\usage{
MixedModels(models, alpha = 1)
}
\arguments{
\item{models}{a named list of DlvmPrior's object}

\item{alpha}{Dirichlet prior parameter over the cluster proportions (default to 1)}
}
\value{
a \code{MixedModels-class} object
}
\description{
An S4 class to represent a mixed clustering models, where sevral models are used to model different datasets. A conditional independence assumption between the view knowing the cluster is made.
}
\details{
The filed name in the models list must match the name of the list use to provide the datasets to cluster together.
}
\section{Functions}{
\itemize{
\item \code{MixedModels}: MixedModels class constructor
}}

\section{Slots}{

\describe{
\item{\code{models}}{a named list of DlvmModels Emission prior such as (\code{\link{GmmPrior}},\code{\link{LcaPrior}},\code{\link{SbmPrior}},...)}
}}

\examples{
MixedModels(models = list(continuous = GmmPrior(), discrete = LcaPrior()))
}
\seealso{
Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DcSbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{DlvmPrior-class}},
\code{\link{Gmm}},
\code{\link{LcaPrior-class}},
\code{\link{MoMPrior-class}},
\code{\link{MoRPrior-class}},
\code{\link{MultSbmPrior-class}},
\code{\link{SbmPrior-class}},
\code{\link{greed}()}
}
\concept{DlvmModels}
