% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm.R
\docType{class}
\name{mmm-class}
\alias{mmm-class}
\title{Mixed mixture model description class}
\description{
An S4 class to represent a Multinomial model model, extends \code{\link{icl_model-class}}.
Such model can be used to cluster a data matrix \eqn{X} with the following generative model :  
\deqn{ \pi \sim Dirichlet(\alpha)}
\deqn{ Z_i  \sim \mathcal{M}(1,\pi)}
\deqn{ \theta_{kv} \sim Dirichlet(\beta)}
....
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter (default to 1)}

\item{\code{beta}}{Dirichlet over vocabulary prior parameter (default to 1)}
}}

\examples{
new("mmm")
new("mmm",alpha=1,beta=1)
}
