% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_classes.R, R/dcsbm.R, R/mm.R, R/mreg.R,
%   R/mvmreg.R, R/sbm.R
\docType{class}
\name{models-classes}
\alias{models-classes}
\alias{icl_model}
\alias{icl_model-class}
\alias{sbm}
\alias{sbm-class}
\alias{mm}
\alias{mm-class}
\alias{dcsbm-class}
\alias{mreg-class}
\alias{mvmreg-class}
\title{Clustering models classes}
\description{
Clustering models classes

icl_model

An S4 class to represent an abstract clustering model
\itemize{
\item slots : \code{name,alpha}
}

dcsbm

An S4 class to represent a stochastick block model that extends \code{icl_model} class.
\itemize{
\item slots : \code{name,alpha,a0,b0}
}

mm

An S4 class to represent a mixture of multinomial also known has mixture of unigrams that extends \code{icl_model} class.
\itemize{
\item slots : \code{name,alpha,beta}
}

mreg

An S4 class to represent a mixture of multinomial also known has mixture of unigrams that extends \code{icl_model} class.
\itemize{
\item slots : \code{name,alpha,reg,a0,b0}
}

mreg

An S4 class to represent a mixture of multinomial also known has mixture of unigrams that extends \code{icl_model} class.
\itemize{
\item slots : \code{name,alpha,beta,N0}
}

sbm

An S4 class to represent a stochastick block model that extends \code{icl_model} class.
\itemize{
\item slots : \code{name,alpha,a0,b0}
}
}
\section{Slots}{

\describe{
\item{\code{name}}{a character vector}

\item{\code{alpha}}{a numeric vector of length 1 which define the parameters of the dirichlet over the cluster proportions (default to 1)}

\item{\code{a0}}{a numeric vector of length 1 which define the parameters of the beta prior over the edges (default to 1)}

\item{\code{b0}}{a numeric vector of length 1 which define the parameters of the beta prior over the non-edges (default to 1)}

\item{\code{beta}}{a numeric vector of length 1 which define the parameters of the beta prior over the counts (default to 1)}

\item{\code{reg}}{a numeric vector of length 1 which define the variance parameter of the normal prior over the regression parameters (default to 0.1)}

\item{\code{a0}}{a numeric vector of length 1 which define the parameter a0 of the inverse gamma over the regression noise variance parameters (default to 1)}

\item{\code{b0}}{a numeric vector of length 1 which define the parameter b0 of the inverse gamma prior over the regression noise variance parameters (default to 1)}

\item{\code{beta}}{a numeric vector of length 1 which define the variance parameter of the normal prior over the regression parameters (default to 0.1)}

\item{\code{N0}}{a numeric vector of length 1 which define the parameter a0 of the inverse gamma over the regression noise variance parameters (default to 1)}

\item{\code{a0}}{a numeric vector of length 1 which define the parameters of the beta prior over the edges (default to 1)}

\item{\code{b0}}{a numeric vector of length 1 which define the parameters of the beta prior over the non-edges (default to 1)}
}}

\examples{
new("dcsbm")
new("mm")
new("mm",alpha=1,beta=1)
new("mreg")
new("mreg",alpha=1,reg=5,a0=0.5,b0=0.5)
new("mreg")
new("mreg",alpha=1,reg=5,a0=0.5,b0=0.5)
new("sbm")
new("sbm",a0=0.5,b0=0.5,alpha=1)
}
