% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mom.R
\docType{class}
\name{MoMPrior-class}
\alias{MoMPrior-class}
\alias{MoM-class}
\alias{MoMPrior}
\alias{MoM}
\title{Mixture of Multinomial Model Prior description class}
\usage{
MoMPrior(beta = 1)

MoM(alpha = 1, beta = 1)
}
\arguments{
\item{beta}{Dirichlet over vocabulary prior parameter (default to 1)}

\item{alpha}{Dirichlet prior parameter over the cluster proportions (default to 1)}
}
\value{
a \code{MoMPrior-class} object

a \code{MoM-class} object
}
\description{
An S4 class to represent a Mixture of Multinomial model.
Such model can be used to cluster a data matrix \eqn{X} with the following generative model :
\deqn{ \pi \sim Dirichlet(\alpha)}
\deqn{ Z_i  \sim \mathcal{M}(1,\pi)}
\deqn{ \theta_{k} \sim Dirichlet(\beta)}
\deqn{ X_{i.}|Z_{ik}=1 \sim \mathcal{M}(L_i,\theta_{k})}
With \eqn{L_i=\sum_d=1^DX_{id}}. These classes mainly store the prior parameters value (\eqn{\alpha,\beta}) of this generative model.
The \code{MoM-class} must be used when fitting a simple Mixture of Multinomials whereas the \code{MoMPrior-class} must be sued when fitting a \code{\link{MixedModels-class}}.
}
\section{Functions}{
\itemize{
\item \code{MoM-class}: MoM class constructor

\item \code{MoMPrior}: MoMPrior class constructor

\item \code{MoM}: MoM class constructor
}}

\section{Slots}{

\describe{
\item{\code{beta}}{Dirichlet over vocabulary prior parameter (default to 1)}

\item{\code{alpha}}{Dirichlet prior parameter over the cluster proportions (default to 1)}
}}

\examples{
MoMPrior()
MoMPrior(beta = 0.5)
MoM()
MoM(beta = 0.5)
}
\seealso{
Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DcSbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{DlvmPrior-class}},
\code{\link{Gmm}},
\code{\link{LcaPrior-class}},
\code{\link{MixedModels-class}},
\code{\link{MoRPrior-class}},
\code{\link{MultSbmPrior-class}},
\code{\link{SbmPrior-class}},
\code{\link{greed}()}
}
\concept{DlvmModels}
