% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mor.R
\docType{class}
\name{MoRPrior-class}
\alias{MoRPrior-class}
\alias{MoR-class}
\alias{MoRPrior}
\alias{MoR}
\title{Multivariate mixture of regression Prior model description class}
\usage{
MoRPrior(formula, tau = 0.01, N0 = NaN, epsilon = NaN)

MoR(formula, alpha = 1, tau = 0.01, N0 = NaN, epsilon = NaN)
}
\arguments{
\item{formula}{a \code{\link{formula}} that describe the linear model to use}

\item{tau}{Prior parameter (inverse variance) default 0.001}

\item{N0}{Prior parameter (default to NaN, in this case N0 will be fixed equal to the number of columns of Y.)}

\item{epsilon}{Covariance matrix prior parameter (default to NaN, in this case epsilon will be fixed to a diagonal variance matrix equal to 0.1 time the variance of the regression residuals with only one cluster.)}

\item{alpha}{Dirichlet prior parameter over the cluster proportions (default to 1)}
}
\value{
a \code{MoRPrior-class} object

a \code{MoR-class} object
}
\description{
An S4 class to represent a multivariate mixture of regression model.
The model follows \href{https://tminka.github.io/papers/minka-linear.pdf}{minka-linear} .
The model corresponds to the following generative model:
\deqn{ \pi \sim Dirichlet(\alpha)}
\deqn{ Z_i  \sim \mathcal{M}(1,\pi)}
\deqn{ V_k \sim \mathcal{W}(\varepsilon^{-1},n_0)}
\deqn{ A_k \sim \mathcal{MN}(0,(V_k)^{-1},\tau X^{t}X)}
\deqn{ Y_{i.}|X_{i.}Z_{ik}=1 \sim \mathcal{N}(A_kx_{i.},V_{k}^{-1})}
with \eqn{\mathcal{W}(\epsilon^{-1},n_0)} the Whishart distribution and \eqn{\mathcal{MN}} the matrix-normal distribution.
The \code{MoR-class} must be used when fitting a simple Mixture of Regression whereas the \code{MoRPrior-class} must be used when fitting a \code{\link{MixedModels-class}}.
}
\section{Functions}{
\itemize{
\item \code{MoR-class}: MoR class constructor

\item \code{MoRPrior}: MoRPrior class constructor

\item \code{MoR}: MoR class constructor
}}

\section{Slots}{

\describe{
\item{\code{formula}}{a \code{\link{formula}} that describe the linear model to use}

\item{\code{tau}}{Prior parameter (inverse variance) default 0.01}

\item{\code{epsilon}}{Covariance matrix prior parameter (default to NaN, in this case epsilon will be fixed to a diagonal variance matrix equal to 0.1 time the variance of the regression residuals with only one cluster.)}

\item{\code{N0}}{Prior parameter (default to NaN, in this case N0 will be fixed equal to the number of columns of Y.)}

\item{\code{alpha}}{Dirichlet prior parameter over the cluster proportions (default to 1)}
}}

\examples{
MoRPrior(y ~ x1 + x2)
MoRPrior(y ~ x1 + x2, N0=100)
MoRPrior(cbind(y1,y2) ~ x1 + x2, N0=100)
MoR(y ~ x1 + x2)
MoR(y ~ x1 + x2, N0=100)
MoR(cbind(y1,y2) ~ x1 + x2, N0=100)
}
\seealso{
Other DlvmModels: 
\code{\link{DcLbmPrior-class}},
\code{\link{DcSbmPrior-class}},
\code{\link{DiagGmmPrior-class}},
\code{\link{DlvmPrior-class}},
\code{\link{GmmPrior-class}},
\code{\link{LcaPrior-class}},
\code{\link{MixedModels-class}},
\code{\link{MoMPrior-class}},
\code{\link{MultSbmPrior-class}},
\code{\link{SbmPrior-class}},
\code{\link{greed}()}
}
\concept{DlvmModels}
