% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{class}
\name{multistarts-class}
\alias{multistarts-class}
\title{Greedy algorithm with multiple start class}
\description{
An S4 class to represent a greedy algorithm  with multiple start (extends \code{\link{alg-class}} class).
}
\section{Slots}{

\describe{
\item{\code{nb_start}}{number of random starts (default to 10)}
}}

