% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{post_probs}
\alias{post_probs}
\title{post_probs}
\usage{
post_probs(model, data, clt)
}
\arguments{
\item{model}{icl_model}

\item{data}{list with clustering data (fileds depend on model type)}

\item{clt}{cluster labels in 1,..,K}
}
\description{
post_probs
}
