% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{post_probs}
\alias{post_probs}
\title{post_probs}
\usage{
post_probs(model, data, clt)
}
\arguments{
\item{model}{icl_model}

\item{clt}{cluster labels {0,...,K-1}}

\item{xp}{sparseMatrix}
}
\description{
post_probs
}
