% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\docType{class}
\name{sbm-class}
\alias{sbm-class}
\title{Stochastic Block Model class}
\description{
An S4 class to represent a Stochastic Block Model, extend \code{\link{icl_model-class}}. 
Such models can be used to cluster graph vertex, and model a square adjacency matrix \eqn{X} with the following generative model :  
\deqn{ \begin{aligned}
\pi &\sim Dirichlet(\alpha)\\
Z_i & \sim \mathcal{M}(1,\pi)\\
\theta_{kl} & \sim Beta(a_0,b_0)\\
X_{ij}|Z_{ik}Z_{jl}=1 & \sim \mathcal{B}(\theta_kl)\\
\end{aligned}}
This class mainly store the prior parameters value \eqn{\alpha,a_0,b_0} of this generative model in the following slots:
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter}

\item{\code{a0}}{Beta prior parameter over links}

\item{\code{b0}}{Beta prior parameter over no-links}
}}

