% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\docType{class}
\name{sbm-class}
\alias{sbm-class}
\title{sbm

An S4 class to represent a stochastick block model that extends \code{icl_model} class.}
\description{
sbm

An S4 class to represent a stochastick block model that extends \code{icl_model} class.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{dirichlet parameter for the prior over cluyster proportions (default to 1)}

\item{\code{a0}}{a numeric vector of length 1 which define the parameters of the beta prior over the edges (default to 1)}

\item{\code{b0}}{a numeric vector of length 1 which define the parameters of the beta prior over the non-edges (default to 1)}
}}

\examples{
new("sbm")
new("sbm",a0=0.5,b0=0.5,alpha=1)
}
