% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\docType{class}
\name{sbm-class}
\alias{sbm-class}
\title{Clustering with a stochastic block model description class}
\description{
An S4 class to represent a a stochastic block model, extend \code{\link{icl_model-class}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the model}

\item{\code{alpha}}{Dirichlet over cluster proportions prior parameter}

\item{\code{a0}}{Beta prior parameter over links}

\item{\code{b0}}{Beta prior parameter over no-links}
}}

