% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\docType{class}
\name{Seed-class}
\alias{Seed-class}
\alias{Seed}
\title{Greedy algorithm with seeded initialization}
\usage{
Seed()
}
\value{
a \code{Seed-class} object
}
\description{
An S4 class to represent a greedy algorithm with initialization from spectral clustering and or k-means (extends \code{\link{Alg-class}} class ).
}
\section{Functions}{
\itemize{
\item \code{Seed()}: Seed algorithm class constructor

}}
\examples{
Seed()
}
