% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{spectral}
\alias{spectral}
\title{spectral
Regularized spectral clustering nips paper 2013}
\usage{
spectral(X, K)
}
\arguments{
\item{X}{An adjacency matrix in sparse format}

\item{K}{Desired number of cluster}
}
\value{
cl Vector of clsuter labels
}
\description{
spectral
Regularized spectral clustering nips paper 2013
}
