% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_classes.R
\name{spectral}
\alias{spectral}
\title{spectral
Regularized spectral clustering}
\usage{
spectral(X, K)
}
\arguments{
\item{X}{An adjacency matrix in sparse format}

\item{K}{Desired number of cluster}
}
\value{
cl Vector of cluster labels
}
\description{
spectral
Regularized spectral clustering
}
\references{
Tai Qin, Karl Rohe. Regularized Spectral Clustering under the Degree-Corrected Stochastic Blockmodel. Nips 2013.
}
